### 2022-05-30 #4420 @Kevin-Mattheus-Moerman
@Xia-Yijie thanks for this submission to JOSS. Can you work on those potentially missing and invalid DOIs? :point_up:. If you fix these you can call `@editorialbot check references` here in a comment to check them again. Thanks.
