### 2022-05-30 #4420 @Kevin-Mattheus-Moerman
Alternative, @jarvist this seems like it might be your cup of tea :tea:, but you are handling many other papers so did not invite you. If you are available let me know as there are other chemistry related papers Richard could work on too if possible. Thanks.
