+++ paper.bib
@@ -6,7 +6,7 @@ volume = {40},
 number = {1},
 pages = {160-168},
 keywords = {Molecular dynamics, Molecular modeling, Enhanced sampling, Machine learning, Computational chemistry},
-doi = {https://doi.org/10.1002/cjoc.202100456},
+doi = {10.1002/cjoc.202100456},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/cjoc.202100456},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/cjoc.202100456},
 abstract = {Comprehensive Summary SPONGE (Simulation Package tOward Next GEneration molecular modeling) is a software package for molecular dynamics (MD) simulation of solution and surface molecular systems. In this version of SPONGE, the all- atom potential energy functions used in AMBER MD packages are used by default and other all-atom/coarse- grained potential energy functions are also supported. SPONGE is designed to extend the timescale being approached in MD simulations by utilizing the latest CUDA- enabled graphical processing units (GPU) and adopting highly efficient enhanced sampling algorithms, such as integrated tempering, selective integrated tempering and enhanced sampling of reactive trajectories. It is highly modular and new algorithms and functions can be incorporated con veniently. Particularly, a specialized Python plugin can be easily used to perform the machine learning MD simulation with MindSpore, TensorFlow, PyTorch or other popular machine learning frameworks. Furthermore, a plugin of Finite-Element Method (FEM) is also available to handle metallic surface systems. All these advanced features increase the power of SPONGE for modeling and simulation of complex chemical and biological systems. What is the most favorite and original chemistry developed in your research group? Our research centers at developing methods and theories to unravel molecular mechanisms of chemical and biological systems. By establishing theoretical models, developing enhanced sampling methods combined with machine learning techniques, we are able to conduct comprehensive thermodynamic and dynamic analyses for these complex systems. How do you get into this specific field? Could you please share some experiences with our readers? I got into theoretical chemistry as a PhD student. My PhD adviser Prof. Rudolph A. Marcus led me into this field and inspired me by his love of science. Enjoy life, always learn new things and be independent in thinking are something I learnt from my advisers (Professors Dalin Yang, Qihe Zhu, Rudy Marcus, and Martin Karplus) and would love to pass to my students. How do you supervise your students? We learn from each other. What is the most important personality for scientific research? Curiosity, passion, and persistence have been of great value to my career. What are your hobbies? What's your favorite book(s)? Reading, Ping-Pong, and jogging. I always enjoy reading history. Who influences you mostly in your life? Too many, family, academic advisors, friends, students, and colleagues.},
@@ -20,7 +20,7 @@ volume = {1-2},
 pages = {19-25},
 year = {2015},
 issn = {2352-7110},
-doi = {https://doi.org/10.1016/j.softx.2015.06.001},
+doi = {10.1016/j.softx.2015.06.001},
 url = {https://www.sciencedirect.com/science/article/pii/S2352711015000059},
 author = {Mark James Abraham and Teemu Murtola and Roland Schulz and Szilárd Páll and Jeremy C. Smith and Berk Hess and Erik Lindahl},
 keywords = {Molecular dynamics, GPU, SIMD, Free energy},
@@ -34,7 +34,7 @@ volume = {271},
 pages = {108171},
 year = {2022},
 issn = {0010-4655},
-doi = {https://doi.org/10.1016/j.cpc.2021.108171},
+doi = {10.1016/j.cpc.2021.108171},
 url = {https://www.sciencedirect.com/science/article/pii/S0010465521002836},
[…]
