### 2022-05-31 #4420 @Xia-Yijie
> @Xia-Yijie thanks for this submission to JOSS. Can you work on those potentially missing and invalid DOIs? ☝️. If you fix these you can call `@editorialbot check references` here in a comment to check them again. Thanks.

Thank you for your reminder!
