### 2022-06-07 a231bd5
update Generalized Born part and corespongding setup and the API document
