### 2022-06-21 cedcfb6
collect the spelling mistakes in API.md
