### 2022-06-21 4367dcc
delete unused comments in BUILD.py
