### 2022-06-21 b78e6d5
replace type(xxx)==yyy with isinstance(xxx,yyy)
