+++ Xponge/IMPOSE.py
@@ -267,7 +267,7 @@ def _link_residue_process_coordinate(molecule, atom1, atom2):
         atom.z = crd[i][2]
         
 def ResidueType_Add(self, other):
-    if type(other) == ResidueType:
+    if isinstance(other, ResidueType):
         new_molecule = Molecule(self.name)
         resA = Residue(self)
         resB = Residue(other)
@@ -283,7 +283,7 @@ def ResidueType_Add(self, other):
             new_molecule.Add_Residue_Link(atom1, atom2)
             _link_residue_process_coordinate(new_molecule, atom1, atom2)            
         return new_molecule
-    elif type(other) == Molecule:
+    elif isinstance(other, Molecule):
         new_molecule = other.deepcopy()
         resA = Residue(self)
         resB = new_molecule.residues[0]
@@ -296,13 +296,13 @@ def ResidueType_Add(self, other):
             new_molecule.Add_Residue_Link(atom1, atom2)
             _link_residue_process_coordinate(new_molecule, atom1, atom2)   
         return new_molecule
-    elif type(other) == type(None):
+    elif other is None:
         return self
     else:
         raise TypeError("unsupported operand type(s) for +: '%s' and '%s'"%(type(self), type(other)))
 
 def Molecule_Add(self, other):
-    if type(other) == ResidueType:
+    if isinstance(other, ResidueType):
         new_molecule = self.deepcopy()
         resA = new_molecule.residues[-1]
         resB = Residue(other)
@@ -315,7 +315,7 @@ def Molecule_Add(self, other):
             new_molecule.Add_Residue_Link(atom1, atom2)
             _link_residue_process_coordinate(new_molecule, atom1, atom2)     
         return new_molecule
-    elif type(other) == Molecule:
+    elif isinstance(other, Molecule):
         new_molecule = self.deepcopy()
         new_molecule2 = other.deepcopy()
         resA = new_molecule.residues[-1]
@@ -330,13 +330,13 @@ def Molecule_Add(self, other):
             new_molecule.Add_Residue_Link(atom1, atom2)
             _link_residue_process_coordinate(new_molecule, atom1, atom2)   
         return new_molecule
-    elif type(other) == type(None):
+    elif other is None::
         return self
     else:
         raise TypeError("unsupported operand type(s) for +: '%s' and '%s'"%(type(self), type(other)))
 
 def iMolecule_Add(self, other):
-    if type(other) == ResidueType:
+    if isinstance(other, ResidueType):
         resA = self.residues[-1]
         resB = Residue(other)
         for atom in other.atoms:
@@ -348,7 +348,7 @@ def iMolecule_Add(self, other):
             self.Add_Residue_Link(atom1, atom2)
             _link_residue_process_coordinate(self, atom1, atom2)  
         return self
-    elif type(other) == Molecule:
+    elif isinstance(other, Molecule):
         new_molecule2 = other.deepcopy()
         resA = self.residues[-1]
         resB = new_molecule2.residues[0]
@@ -362,7 +362,7 @@ def iMolecule_Add(self, other):
             self.Add_Residue_Link(atom1, atom2)
             _link_residue_process_coordinate(self, atom1, atom2)   
         return self
-    elif type(other) == type(None):
+    elif other is None:
         return self
     else:
         raise TypeError("unsupported operand type(s) for +: '%s' and '%s'"%(type(self), type(other)))
@@ -381,7 +381,7 @@ def Muls(self, other):
         raise TypeError("unsupported operand type(s) for *: '%s' and '%s'"%(type(self), type(other)))
 
 def iMuls(self, other):
-    if type(other) == int:
+    if isinstance(other,int):
         assert other >= 1
         for i in range(other - 1):
[…]
