### 2022-06-21 cfcb089
change the default function values from {}, [] to None
