+++ Xponge/__init__.py
@@ -90,8 +90,9 @@ class Type:
     types_different_name = {}
 
     @classmethod
-    def Add_Property(cls, parm_fmt, parm_default = {}):
-        cls.parameters.update(parm_fmt)
+    def Add_Property(cls, parm_fmt, parm_default = None):
+        if parm_default is not None:
+            cls.parameters.update(parm_fmt)
         for _type in cls.types.values():
             _type.contents.update({key:parm_default.get(key, None) for key in parm_fmt.keys()})
 
