### 2022-06-21 a57f1ab
fix 3 bare except. 1 bare except in resp still need to be fixed
