### 2022-06-21 c3c6713
fix the bugs when cleaning the code
