### 2022-06-21 91da865
make some long functions shorter a little bit
