+++ Xponge/__main__.py
@@ -1,15 +1,11 @@
-def main():
-    from . import tools
-    import argparse
-
-    parser = argparse.ArgumentParser(prog = "Xponge")
-
-    subparsers = parser.add_subparsers(help = "subcommands", description = "Tools for SPONGE. Use Xponge XXX -h for the help of tool 'XXX'.")
-
+from . import tools
+    
+def _mytest(subparsers):
     mytest = subparsers.add_parser("test", help = "test the basic function of Xponge")
     mytest.add_argument("-o", metavar = "test", default="test", help="the prefix for the output files")
     mytest.set_defaults(func = tools.test)
-    
+
+def _dat2nc(subparsers):
     dat2nc = subparsers.add_parser("dat2nc", help = "convert a traj file from .dat to .nc")
 
     dat2nc.add_argument("-n",required=True,type=int, help="the number of atoms in the traj file",metavar="atom_numbers")
@@ -23,16 +19,19 @@ def main():
     dat2nc.add_argument("--title",default="DEFAULT", help="the title of the traj.")
     dat2nc.set_defaults(func = tools.dat2nc)
 
+def _gro2crd(subparsers):
     gro2crd = subparsers.add_parser("gro2crd", help = "convert a coordinate file from gromacs to SPONGE")
     gro2crd.add_argument("-i",required=True, help="the name of the .gro file.")
     gro2crd.add_argument("-o", required=True, help="the name of the SPONGE coordinate file.")
     gro2crd.set_defaults(func = tools.gro2crd)
 
+def _nc2rst7(subparsers):
     nc2rst7 = subparsers.add_parser("nc2rst7", help = "convert a rst7 file from .nc to .rst7")
     nc2rst7.add_argument("-rst7",required=True, help="the name of the .rst7 restart file.")
     nc2rst7.add_argument("-nc", required=True, help="the name of the .nc restart file.")
     nc2rst7.set_defaults(func = tools.nc2rst7)
 
+def _maskgen(subparsers):
     maskgen = subparsers.add_parser("maskgen", help = "use VMD to generate a file to record the atom indexes of the corresponding mask")
     maskgen.add_argument("-p",required=True, help="the topology file")
     maskgen.add_argument("-c", help="the coordinate file")
@@ -40,6 +39,7 @@ def main():
     maskgen.add_argument("--vmd", metavar = "vmd", default="vmd", help="the command to start vmd")
     maskgen.set_defaults(func = tools.maskgen)
 
+def _exgen(subparsers):
     exgen = subparsers.add_parser("exgen", help = 'process bond-like, angle-like, dihedral-like files to get the atoms to exclude')
     exgen.add_argument('-n', type = int, required = True, help='the atom numbers')
     exgen.add_argument('-o', required = True, help='output exclude file name')
@@ -50,6 +50,7 @@ def main():
     exgen.add_argument('-e', '--exclude', nargs='+', help='exclude-like input files: add the information of another exclude file')
     exgen.set_defaults(func = tools.exgen)
 
+def _dat1frame(subparsers):
     dat1frame = subparsers.add_parser("dat1frame", help = 'extract 1 frame from .dat traj file')
     dat1frame.add_argument('-n', type = int, required = True, help='the atom numbers')
     dat1frame.add_argument('-frame', type = int, required = True, help='the frame to extract')
@@ -57,7 +58,8 @@ def main():
     dat1frame.add_argument('-box', required = True, help='box traj file name')
     dat1frame.add_argument('-dat', required = True, help='dat traj file name')
     dat1frame.set_defaults(func = tools.dat1frame)
-
+    
+def _crd2rst7(subparsers):
[…]
