### 2022-06-21 1fcf4f9
fix bugs when cleaning the codes
