### 2022-06-21 78acb44
reduce the cyclomatic comlexity of exgen
