### 2022-06-22 22d3762
clean mol2rfe function
