+++ Xponge/tools/__init__.py
@@ -418,6 +418,116 @@ def crd2pdb(args):
     with open(args.o, "w") as f:
         f.write(towrite)    
 
+def mol2rfe_build(args, merged_from, merged_to):
+    import os 
+    import Xponge
+    import Xponge.forcefield.SPECIAL.FEP as FEP
+    import Xponge.forcefield.SPECIAL.MIN_BONDED as MIN
+
+    if "build" in args.do:
+        print("\nBUILDING TOPOLOGY\n")
+        FEP.Save_Soft_Core_LJ()
+
+        for i in range(args.nl + 1):
+            if os.path.exists("%d"%i):
+                os.system("rm -rf %d"%i)
+            os.mkdir("%d"%i)
+            tt = FEP.Merge_Force_Field(merged_from, merged_to,  i / args.nl)
+            if i == 0:
+                MIN.Save_Min_Bonded_Parameters()
+            elif i == 1:
+                MIN.Do_Not_Save_Min_Bonded_Parameters()
+            Xponge.BUILD.Save_SPONGE_Input(tt, "%d/%s"%(i, args.temp))
+
+def mol2rfe_min(args):
+    import os 
+    import Xponge
+    
+    if "min" in args.do:
+        for i in range(args.nl + 1):
+            if os.path.exists("%d/min"%i):
+                os.system("rm -rf %d/min"%i)
+            os.mkdir("%d/min"%i)
+            if i != 0:
+                cif = "-coordinate_in_file {1}/min/{0}_coordinate.txt".format(args.temp, i-1)
+            else:
+                cif = ""
+                os.system("{0} -mode minimization -minimization_dynamic_dt 1 -default_in_file_prefix {2}/{1} -mdinfo {2}/min/{1}.mdinfo -mdout {2}/min/{1}.mdout -rst {2}/min/{1} -crd {2}/min/{1}.dat -box {2}/min/{1}.box -step_limit {5} {4} -mass_in_file {2}/{1}_fake_mass.txt -lambda_lj {3}".format(args.sponge, args.temp, i, i / args.nl, cif, args.m1steps[0]))
+                cif = "-coordinate_in_file {1}/min/{0}_coordinate.txt".format(args.temp, i)
+                os.system("{0} -mode minimization -dt 1e-7 -default_in_file_prefix {2}/{1} -mdinfo {2}/min/{1}.mdinfo -mdout {2}/min/{1}.mdout -rst {2}/min/{1} -crd {2}/min/{1}.dat -box {2}/min/{1}.box -step_limit {5}  {4} -mass_in_file {2}/{1}_fake_mass.txt -lambda_lj {3}".format(args.sponge, args.temp, i, i / args.nl, cif, args.m1steps[1]))
+                os.system("{0} -mode minimization -dt 1e-6 -default_in_file_prefix {2}/{1} -mdinfo {2}/min/{1}.mdinfo -mdout {2}/min/{1}.mdout -rst {2}/min/{1} -crd {2}/min/{1}.dat -box {2}/min/{1}.box -step_limit {5}  {4} -mass_in_file {2}/{1}_fake_mass.txt -lambda_lj {3}".format(args.sponge, args.temp, i, i / args.nl, cif, args.m1steps[2]))
[…]
