### 2022-06-22 c4947de
clean up write_LJ
