+++ Xponge/forcefield/SPECIAL/FEP.py
@@ -252,140 +252,20 @@ def Save_Soft_Core_LJ():
             if atom.LJtypeB not in LJtypemapB.keys():
                 LJtypemapB[atom.LJtypeB] = len(LJtypesB)
                 LJtypesB.append(atom.LJtypeB)
+        
+        As, Bs = LJ.find_AB_LJ(LJtypes)
+        AsA, BsB = LJ.find_AB_LJ(LJtypesB)
 
-        As = []
-        Bs = []
-        for i in range(len(LJtypes)):
-            LJ_i = LJType.types[LJtypes[i] + "-" + LJtypes[i]]
-            for j in range(len(LJtypes)):
-                LJ_j = LJType.types[LJtypes[j] + "-" + LJtypes[j]]
-                finded = False
-                findnames = [LJtypes[i] + "-" + LJtypes[j], LJtypes[j] + "-" + LJtypes[i]]
-                for findname in findnames:
-                    if findname in LJType.types.keys():
-                        finded = True
-                        LJ_ij = LJType.types[findname]
-                        As.append(LJType.combining_method_A(LJ_ij.epsilon, LJ_ij.rmin, LJ_ij.epsilon, LJ_ij.rmin))
-                        Bs.append(LJType.combining_method_B(LJ_ij.epsilon, LJ_ij.rmin, LJ_ij.epsilon, LJ_ij.rmin))
-                        break
-                if not finded:
-                    As.append(LJType.combining_method_A(LJ_i.epsilon, LJ_i.rmin, LJ_j.epsilon, LJ_j.rmin))
-                    Bs.append(LJType.combining_method_B(LJ_i.epsilon, LJ_i.rmin, LJ_j.epsilon, LJ_j.rmin))
+        checks = LJ.get_checks(LJtypes, As, Bs)
+        same_type = LJ.judge_same_type(LJtypes, checks)
+        real_LJtypes = LJ.get_real_LJ(LJtypes, same_type)
+        real_As, real_Bs = LJ.find_AB_LJ(real_LJtypes)
 
-        AsB = []
-        BsB = []
-        for i in range(len(LJtypesB)):
-            LJ_i = LJType.types[LJtypesB[i] + "-" + LJtypesB[i]]
-            for j in range(len(LJtypesB)):
-                LJ_j = LJType.types[LJtypesB[j] + "-" + LJtypesB[j]]
-                finded = False
-                findnames = [LJtypesB[i] + "-" + LJtypesB[j], LJtypesB[j] + "-" + LJtypesB[i]]
-                for findname in findnames:
-                    if findname in LJType.types.keys():
-                        finded = True
-                        LJ_ij = LJType.types[findname]
-                        AsB.append(LJType.combining_method_A(LJ_ij.epsilon, LJ_ij.rmin, LJ_ij.epsilon, LJ_ij.rmin))
-                        BsB.append(LJType.combining_method_B(LJ_ij.epsilon, LJ_ij.rmin, LJ_ij.epsilon, LJ_ij.rmin))
-                        break
-                if not finded:
-                    AsB.append(LJType.combining_method_A(LJ_i.epsilon, LJ_i.rmin, LJ_j.epsilon, LJ_j.rmin))
-                    BsB.append(LJType.combining_method_B(LJ_i.epsilon, LJ_i.rmin, LJ_j.epsilon, LJ_j.rmin))
-
-        checks = {}
-        count = 0
-        for i in range(len(LJtypes)):
-            check_string_A = ""
-            check_string_B = ""
-            for j in range(len(LJtypes)):
-                check_string_A += "%16.7e" % As[count] + " "
-                check_string_B += "%16.7e" % Bs[count] + " "
-                count += 1
-
-            checks[i] = check_string_A + check_string_B
-
[…]
