### 2022-06-22 1fa615c
clean up FEP write_LJ
