### 2022-06-22 560cf8d
clean up write_LJ
