### 2022-06-22 8afeead
clean up merge dual topology
