### 2022-06-22 c1d7ecf
clean up merge dual topology
