### 2022-06-22 2a81a37
clean up merge dual topology
