### 2022-06-22 364a7b3
clean up merge dual topology
