### 2022-06-22 d3c3f58
clean up merge dual topology
