+++ Xponge/forcefield/SPECIAL/FEP.py
@@ -349,9 +349,7 @@ def Get_Free_Molecule(molA, perturbing_residues, intra_FEP=False):
 
     return molB
 
-def _get_ResidueAB(ResidueTypeA, ResidueTypeB, ResidueA, forcopy, matchmap, matchA, matchB):
-    restypeAB = ResidueTypeA.deepcopy(ResidueTypeA.name + "_" + ResidueTypeB.name, forcopy)
-
+def _get_extra_atoms_and_RBmap(restypeAB, ResidueTypeA, ResidueTypeB, ResidueA, forcopy, matchmap, matchA, matchB):
     extraA = []
     extraB = []
     RBmap = {value: key for key, value in matchmap.items()}
@@ -376,6 +374,20 @@ def _get_ResidueAB(ResidueTypeA, ResidueTypeB, ResidueA, forcopy, matchmap, matc
             extraB[-1].subsys = 2
         else:
             ResidueTypeB.atoms[i].copied[forcopy] = restypeAB.atoms[matchmap[i]]
+    return extraA, extraB, RBmap
+
+def _link_restypeB_atoms(ResidueTypeB, forcopy, matchmap):
+    for atom in ResidueTypeB.atoms:
+        for key, linked_atoms in atom.copied[forcopy].linked_atoms.items():
+            for aton in atom.linked_atoms.get(key, []):
+                if not (ResidueTypeB._atom2index[aton] in matchmap.keys()
+                        and ResidueTypeB._atom2index[atom] in matchmap.keys()):
+                    atom.copied[forcopy].Link_Atom(key, aton.copied[forcopy])
+
+def _get_ResidueAB(ResidueTypeA, ResidueTypeB, ResidueA, forcopy, matchmap, matchA, matchB):
+    restypeAB = ResidueTypeA.deepcopy(ResidueTypeA.name + "_" + ResidueTypeB.name, forcopy)
+
+    extraA, extraB, RBmap = _get_extra_atoms_and_RBmap(restypeAB, ResidueTypeA, ResidueTypeB, ResidueA, forcopy, matchmap, matchA, matchB)
 
     for atomi in extraA:
         for atomj in extraB:
@@ -395,15 +407,8 @@ def _get_ResidueAB(ResidueTypeA, ResidueTypeB, ResidueA, forcopy, matchmap, matc
                     break
             if tocopy:
                 restypeAB.Add_Bonded_Force(bond_entity.deepcopy(forcopy))
-    
                           
-
-    for atom in ResidueTypeB.atoms:
-        for key, linked_atoms in atom.copied[forcopy].linked_atoms.items():
-            for aton in atom.linked_atoms.get(key, []):
-                if not (ResidueTypeB._atom2index[aton] in matchmap.keys()
-                        and ResidueTypeB._atom2index[atom] in matchmap.keys()):
-                    atom.copied[forcopy].Link_Atom(key, aton.copied[forcopy])
+    _link_restypeB_atoms(ResidueTypeB, forcopy, matchmap)
     
     for atom in ResidueTypeA.atoms:
         atom.copied.pop(forcopy)
