+++ Xponge/LOAD.py
@@ -400,6 +400,24 @@ def parmdat(filename):
         
 sys.modules['__main__'].__dict__["loadparmdat"] = parmdat 
 
+def _read_parm7_someflag(length, word_process, offset, ):
+    tempvar = []
+    line = f.readline()
+    while line.startswith("%"):
+        line = f.readline()
+    while 1:
+        i = 0
+        while i + offset < len(line):
+            word = line[i:i+offset].strip()
+            if word:
+                tempvar.append(word_process(word))
+            else:
+                break
+            i += offset
+        if len(tempvar) == length:
+            break
+        line = f.readline()
+
 def parm7(filename):
     import Xponge.forcefield.AMBER
     bonds = {}
@@ -441,68 +459,11 @@ def parm7(filename):
                 angle_type_numbers = int(line[48:56])
                 dihedral_type_numbers = int(line[56:64])
             elif line.startswith("%FLAG DIHEDRAL_FORCE_CONSTANT"):
-                dihedral_k = []
-                tempvar = dihedral_k
-                length = dihedral_type_numbers
-                word_process = lambda x:float(x)
-                offset = 16
-                line = f.readline()
-                while line.startswith("%"):
-                    line = f.readline()
-                while 1:
-                    i = 0
-                    while i + offset < len(line):
-                        word = line[i:i+offset].strip()
-                        if word:
-                            tempvar.append(word_process(word))
-                        else:
-                            break
-                        i += offset
-                    if len(tempvar) == length:
-                        break
-                    line = f.readline()
+                dihedral_k = _read_parm7_someflag(dihedral_type_numbers, lambda x:float(x), 16, f)
             elif line.startswith("%FLAG DIHEDRAL_PERIODICITY"):
-                dihedral_n = []
-                tempvar = dihedral_n
-                length = dihedral_type_numbers
-                word_process = lambda x:float(x)
-                offset = 16
-                line = f.readline()
-                while line.startswith("%"):
-                    line = f.readline()
-                while 1:
-                    i = 0
-                    while i + offset < len(line):
-                        word = line[i:i+offset].strip()
-                        if word:
-                            tempvar.append(word_process(word))
-                        else:
-                            break
-                        i += offset
-                    if len(tempvar) == length:
-                        break
-                    line = f.readline()
+                dihedral_n = _read_parm7_someflag(dihedral_type_numbers, lambda x:float(x), 16, f)
             elif line.startswith("%FLAG DIHEDRAL_PHASE"):
-                dihedral_phase = []
-                tempvar = dihedral_phase
-                length = dihedral_type_numbers
-                word_process = lambda x:float(x)
-                offset = 16
-                line = f.readline()
-                while line.startswith("%"):
-                    line = f.readline()
-                while 1:
-                    i = 0
-                    while i + offset < len(line):
-                        word = line[i:i+offset].strip()
-                        if word:
-                            tempvar.append(word_process(word))
-                        else:
-                            break
-                        i += offset
-                    if len(tempvar) == length:
-                        break
-                    line = f.readline()
+                dihedral_phase = _read_parm7_someflag(dihedral_type_numbers, lambda x:float(x), 16, f)
             elif line.startswith("%FLAG DIHEDRALS_INC_HYDROGEN"):
                 tempvar = []
                 while line.startswith("%"):
@@ -569,47 +530,9 @@ def parm7(filename):
                                 nb14s[tempvar[i+4] - 1] = [[tempvar[i] // 3, tempvar[i+3] // 3]]
 
             elif line.startswith("%FLAG ANGLE_FORCE_CONSTANT"):
-                angle_k = []
-                tempvar = angle_k
-                length = angle_type_numbers
-                word_process = lambda x:float(x)
-                offset = 16
[…]
