+++ Xponge/LOAD.py
@@ -400,7 +400,7 @@ def parmdat(filename):
         
 sys.modules['__main__'].__dict__["loadparmdat"] = parmdat 
 
-def _read_parm7_someflag(length, word_process, offset, ):
+def _read_parm7_someflag(length, word_process, offset, f):
     tempvar = []
     line = f.readline()
     while line.startswith("%"):
@@ -417,9 +417,35 @@ def _read_parm7_someflag(length, word_process, offset, ):
         if len(tempvar) == length:
             break
         line = f.readline()
+    return tempvar
 
-def parm7(filename):
-    import Xponge.forcefield.AMBER
+def _read_parm7_build(mol, BONDTYPE, bonds, parameter_head, parameter_lists, add_connectivity = False):
+    bond_names = {}
+    newBONDs = ""
+    newBONDs += parameter_head
+    for i in bonds.keys():
+        bond_names[i] = "-".join([mol.atoms[atype].type.name for atype in bonds[i][0]])
+        newBONDs += bond_names[i]
+        for fmt, parm_list in parameter_lists:
+            newBONDs += " " + fmt%parm_list[i]
+        newBONDs += "\n"
+    BONDTYPE.New_From_String(newBONDs)
+    
+    for typei, bondi in bonds.items():
+        for atom_indexes in bondi:
+            atoms = [mol.atoms[i] for i in atom_indexes ]
+            residues = set([atom.residue for atom in atoms])
+            if len(residues) == 1:
+                residue = atoms[0].residue
+                resatoms = [residue.type._name2atom[atom.name] for atom in atoms]
+                residue.type.Add_Bonded_Force(BONDTYPE.entity(resatoms, BONDTYPE.types_different_name[bond_names[typei]]))
+                residue.Add_Bonded_Force(BONDTYPE.entity(atoms, BONDTYPE.types_different_name[bond_names[typei]]))
+                if add_connectivity:
+                    residue.Add_Connectivity(atoms[0], atoms[-1])
+            else:
+                raise NotImplementedError
+                
+def _read_parm7_read(filename):   
     bonds = {}
     angles = {}
     propers = {}
@@ -465,21 +491,7 @@ def parm7(filename):
             elif line.startswith("%FLAG DIHEDRAL_PHASE"):
                 dihedral_phase = _read_parm7_someflag(dihedral_type_numbers, lambda x:float(x), 16, f)
             elif line.startswith("%FLAG DIHEDRALS_INC_HYDROGEN"):
-                tempvar = []
-                while line.startswith("%"):
-                    line = f.readline()
-                while 1:
-                    i = 0
-                    while i + 8 < len(line):
-                        word = line[i:i+8].strip()
-                        if word:
-                            tempvar.append(int(word))
-                        else:
-                            break
-                        i += 8
-                    if len(tempvar) == 5 * dihedral_numbers_H:
-                        break
-                    line = f.readline()
+                tempvar = _read_parm7_someflag(5 * dihedral_numbers_H, lambda x:int(x), 8, f)
                 for i in range(0,len(tempvar),5):
                     if tempvar[i+3] < 0:
                         if tempvar[i+4] - 1 in impropers.keys():
@@ -497,21 +509,7 @@ def parm7(filename):
                             else:
                                 nb14s[tempvar[i+4] - 1] = [[tempvar[i] // 3, tempvar[i+3] // 3]]
             elif line.startswith("%FLAG DIHEDRALS_WITHOUT_HYDROGEN"):
-                tempvar = []
-                while line.startswith("%"):
[…]
