### 2022-07-05 9103849
clean the gaff temp functions
