### 2022-07-05 ebb128b
clean RESP calculation
