+++ Xponge/assign/__init__.py
@@ -23,6 +23,73 @@ class Judge_Rule:
 
 
 class _RING():
+    @staticmethod
+    def Add_Rings_Basic_Marker(assign, rings):
+        for ring in rings:
+            for atom in ring.atoms:
+                assign.Add_Atom_Marker(atom, "RG")
+                assign.Add_Atom_Marker(atom, "RG%d" % len(ring.atoms))
+        
+    @staticmethod
+    def Check_Rings_Type(assign, rings):
+        for ring in rings:
+            ring.check_pure_aromatic(assign)
+            ring.check_pure_aliphatic_and_planar(assign)
+            ring.check_out_plane_double_bond(assign)
+
+            if not ring.is_pure_aromatic_ring:
+                for atom in ring.atoms:
+                    if ring.is_pure_aliphatic_ring:
+                        assign.Add_Atom_Marker(atom, "AR5")
+                    elif ring.is_planar_ring:
+                        if ring.out_plane_double_bond:
+                            assign.Add_Atom_Marker(atom, "AR3")
+                        else:
+                            assign.Add_Atom_Marker(atom, "AR2")
+                    else:
+                        assign.Add_Atom_Marker(atom, "AR4")
+        
+    @staticmethod
+    def Get_Rings(assign):
+        current_path = []
+        current_path_sons = {}
+        current_work = []
+        current_path_father = {}
+        have_found_rings = set([])
+        for atom0 in range(len(assign.atoms)):
+            current_path.append(atom0)
+            current_work.extend([[atom, atom0] for atom in assign.bonds[atom0].keys()])
+            current_path_sons[atom0] = len(assign.bonds[atom0])
+            current_path_father = []
+            while current_path:
+                work_atom, from_atom = current_work.pop()
+                current_path.append(work_atom)
+                current_path_father.append(from_atom)
+                bond_atom = []
+                for atom in assign.bonds[work_atom].keys():
+                    if atom != from_atom:
+                        try:
+                            index = current_path.index(atom)
+                            have_found_rings.add(_RING(current_path[index:]))
+                        except ValueError:
+                            bond_atom.append([atom, work_atom])
+
+                if len(current_path) < 9:
+                    current_path_sons[work_atom] = len(bond_atom)
+                    current_work.extend(bond_atom)
+
+                else:
+                    current_path_sons[work_atom] = 0
+
+                for atom in current_path[::-1]:
+                    if current_path_sons[atom] == 0:
+                        pop_atom = current_path.pop()
+                        current_path_sons.pop(pop_atom)
+                        if current_path_father:
+                            father_atom = current_path_father.pop()
+                            current_path_sons[father_atom] -= 1
+        return have_found_rings
+        
     def __repr__(self):
         return self.tohash
 
@@ -44,6 +111,51 @@ class _RING():
     def get_3_neighbors(self):
         for i in range(len(self.atoms)):
             yield self.atoms[i - 2], self.atoms[i - 1], self.atoms[i]
+    
+    def check_pure_aromatic(self, assign):
+        if len(self.atoms) == 6:
+            self.is_pure_aromatic_ring = True
+            for atom in self.atoms:
+                if not assign.Atom_Judge(atom, "C3") and not assign.Atom_Judge(atom, "N2") and not assign.Atom_Judge(atom, "N3"):
+                    self.is_pure_aromatic_ring = False
+                    break
+                if assign.Atom_Judge(atom, "N3"):
+                    temp = 0
+                    for bonded_atom, bond_order in assign.bonds[atom].items():
+                        temp += bond_order
+                    if temp == 3:
[…]
