+++ Xponge/assign/RESP.py
@@ -1,6 +1,6 @@
 try:
     import pyscf
-except:
+except ModuleNotFoundError:
     raise Exception("To calculate RESP charge, 'pyscf' package needed. Maybe you need 'pip install pyscf'")
 
 
@@ -52,11 +52,7 @@ def force_equivalence_q(q, extra_equivalence):
         q[eq_group] = q_mean
     return q
 
-#Pay Attention To !!!UNIT!!!
-def RESP_Fit(Assign, basis = "6-31g*", opt = False, opt_params = None, charge = 0, spin = 0, extra_equivalence = None, grid_density = 6, grid_cell_layer = 4, 
-    radius = None, a1 = 0.0005, a2 = 0.001, two_stage = True, only_ESP  = False):
-    if extra_equivalence is None:
-        extra_equivalence = []
+def _get_pyscf_mol(Assign, basis, charge, spin, opt):
     from pyscf import gto, scf
     mols = ""
     for i, atom in enumerate(Assign.atoms):
@@ -78,8 +74,141 @@ def RESP_Fit(Assign, basis = "6-31g*", opt = False, opt_params = None, charge = 
         Assign.coordinate = mol.atom_coords() * 0.52918
     
     fun.run()
+    return mol, fun
+
+def _resp_scf_kernel(mol, Assign, a, b, A, A0, B, q):
+    step = 0
+    while step == 0 or np.max(np.abs(q - q_last_step)) > 1e-4:
+        step += 1
+        q_last_step = q
+        for i in range(mol.natm):
+            if Assign.atoms[i] != "H":
+                A[i][i] = A0[i][i] + a / np.sqrt(q_last_step[i] *q_last_step[i] + b * b)
+
+        Ainv = np.linalg.inv(A)
+        q = np.dot(Ainv, B)
+        q = q[:-1]
+        
+    return q
+
+def _find_tofit_second(mol, Assign):
+    tofit_second = []
+    fit_group = {i : -1 for i in range(mol.natm)}
+    sublength = 0
+    for i in range(mol.natm):
+        if Assign.Atom_Judge(i, "C4"):
+            fit_group[i] = len(tofit_second)
+            tofit_second.append([i])
+            temp = []
+            for j in Assign.bonds[i].keys():
+                if Assign.atoms[j] == "H":
+                    temp.append(j)
+            if temp:
+                for j in temp:
+                    fit_group[j] = len(tofit_second)
+                tofit_second.append(temp)
+                sublength += len(temp) - 1
+                    
+        if Assign.Atom_Judge(i, "C3"):
+            temp = []
+            for j in Assign.bonds[i].keys():
+                if Assign.atoms[j] == "H":
+                    temp.append(j)
+            if len(temp) == 2:
+                fit_group[i] = len(tofit_second)
+                tofit_second.append([i])
+                for j in temp:
+                    fit_group[j] = len(tofit_second)
+                tofit_second.append(temp)
+                sublength += 1
+    return tofit_second, fit_group, sublength
+
+def _correct_extra_equivalence(tofit_second, fit_group, sublength, extra_equivalence):
+    if extra_equivalence:
+        equi_group = [set() for i in extra_equivalence]    
+        for i, eq in enumerate(extra_equivalence):
+            for eq_atom in eq:
+                if fit_group[eq_atom] != -1:
+                    equi_group[i].add(fit_group[eq_atom])
+            equi_group[i] = list(equi_group[i])
+            equi_group[i].sort()
+            
+        all_groups = set()
+        new_all_groups = set()
+        for atom in range(len(Assign.atoms)):
+            all_groups.add(fit_group[atom])
+        all_groups = list(all_groups)
+        all_groups.sort()
+        
+        
+        group_map = {i:i for i in all_groups}
+        for eq in equi_group:
+            for group in eq:
[…]
