### 2022-07-06 eb03d64
clean up loadmol2
