### 2022-07-08 22a3b9e
modify FEP build and min
