+++ Xponge/tools/__init__.py
@@ -448,21 +448,21 @@ def mol2rfe_min(args):
             if os.path.exists("%d/min"%i):
                 os.system("rm -rf %d/min"%i)
             os.mkdir("%d/min"%i)
-            if i != 0:
+            if i != 0 and args.mlast:
                 cif = "-coordinate_in_file {1}/min/{0}_coordinate.txt".format(args.temp, i-1)
             else:
                 cif = ""
-                os.system("{0} -mode minimization -minimization_dynamic_dt 1 -default_in_file_prefix {2}/{1} -mdinfo {2}/min/{1}.mdinfo -mdout {2}/min/{1}.mdout -rst {2}/min/{1} -crd {2}/min/{1}.dat -box {2}/min/{1}.box -step_limit {5} {4} -mass_in_file {2}/{1}_fake_mass.txt -lambda_lj {3}".format(args.sponge, args.temp, i, i / args.nl, cif, args.m1steps[0]))
+                os.system("{0} -mode minimization -minimization_dynamic_dt 1 -default_in_file_prefix {2}/{1} -mdinfo {2}/min/{1}.mdinfo -mdout {2}/min/{1}.mdout -rst {2}/min/{1} -crd {2}/min/{1}.dat -box {2}/min/{1}.box -step_limit {5} {4} -mass_in_file 0/{1}_fake_mass.txt -lambda_lj {3}".format(args.sponge, args.temp, i, i / args.nl, cif, args.m1steps[0]))
                 cif = "-coordinate_in_file {1}/min/{0}_coordinate.txt".format(args.temp, i)
-                os.system("{0} -mode minimization -dt 1e-7 -default_in_file_prefix {2}/{1} -mdinfo {2}/min/{1}.mdinfo -mdout {2}/min/{1}.mdout -rst {2}/min/{1} -crd {2}/min/{1}.dat -box {2}/min/{1}.box -step_limit {5}  {4} -mass_in_file {2}/{1}_fake_mass.txt -lambda_lj {3}".format(args.sponge, args.temp, i, i / args.nl, cif, args.m1steps[1]))
-                os.system("{0} -mode minimization -dt 1e-6 -default_in_file_prefix {2}/{1} -mdinfo {2}/min/{1}.mdinfo -mdout {2}/min/{1}.mdout -rst {2}/min/{1} -crd {2}/min/{1}.dat -box {2}/min/{1}.box -step_limit {5}  {4} -mass_in_file {2}/{1}_fake_mass.txt -lambda_lj {3}".format(args.sponge, args.temp, i, i / args.nl, cif, args.m1steps[2]))
-                os.system("{0} -mode minimization -dt 1e-5 -default_in_file_prefix {2}/{1} -mdinfo {2}/min/{1}.mdinfo -mdout {2}/min/{1}.mdout -rst {2}/min/{1} -crd {2}/min/{1}.dat -box {2}/min/{1}.box -step_limit {5}  {4} -mass_in_file {2}/{1}_fake_mass.txt -lambda_lj {3}".format(args.sponge, args.temp, i, i / args.nl, cif, args.m1steps[3]))
[…]
