+++ Xponge/LOAD.py
@@ -212,6 +212,32 @@ sys.modules['__main__'].__dict__["loadpdb"] = pdb
 ##########################################################################
 #AMBER Format
 ##########################################################################
+def _frcmod_cmap(line, cmap, temp_cmp, cmap_flag):
+    if line.startswith("%FLAG"):
+        if "CMAP_COUNT" in line:
+            if temp_cmp:
+                for res in temp_cmp["residues"]:
+                    cmap[res] = temp_cmp["info"]
+            temp_cmp = {"residues":[], "info": {"resolution":24, "count":int(line.split()[-1]), "parameters": []}}
+            cmap_flag = "CMAP_COUNT"
+        elif "CMAP_RESOLUTION" in line:
+            temp_cmp["info"]["resolution"] = int(line.split()[-1])
+            cmap_flag = "CMAP_RESOLUTION"
+        elif  "CMAP_RESLIST" in line:
+            cmap_flag = "CMAP_RESLIST"
+        elif "CMAP_TITLE" in line:
+            cmap_flag = "CMAP_TITLE"
+        elif "CMAP_PARAMETER" in line:
+            cmap_flag = "CMAP_PARAMETER"
+    elif cmap_flag == "CMAP_RESLIST":
+        temp_cmp["residues"].extend(line.split())
+    elif cmap_flag == "CMAP_PARAMETER":
+        temp_cmp["info"]["parameters"].extend([float(x) for x in line.split()])
+    return temp_cmp, cmap_flag
+
+def _frcmod_atoms_words(line, N):
+    return [ word.strip() for word in line[:N].split("-") ], line[N:].split()
+    
 def frcmod(filename, nbtype = "RE"):
     with open(filename) as f:
         f.readline()
@@ -224,7 +250,7 @@ def frcmod(filename, nbtype = "RE"):
         impropers = "name  k[kcal/mol]    phi0[degree]    periodicity\n"
         cmap = {}
         cmap_flag = None
-        temp_cmp = {}
+        temp_cmp = {"residues":[]}
         if nbtype == "SK":
             raise NotImplementedError
         elif nbtype == "AC":
@@ -241,62 +267,46 @@ def frcmod(filename, nbtype = "RE"):
             elif flag[:4] == "MASS":
                 atom_types[words[0]] = words[1]
             elif flag[:4] == "BOND":
-                atoms = [ word.strip() for word in line[:5].split("-") ]
-                words = line[5:].split()
+                atoms, words = _frcmod_atoms_words(line, 5)
                 bonds += "-".join(atoms) + "\t" + words[0] + "\t" + words[1] + "\n"
             elif flag[:4] == "ANGL":
-                atoms = [ word.strip() for word in line[:8].split("-") ]
-                words = line[8:].split()
+                atoms, words = _frcmod_atoms_words(line, 8)
                 angles += "-".join(atoms) + "\t" + words[0] + "\t" + words[1] + "\n"
             elif flag[:4] == "DIHE":
-                atoms = [ word.strip() for word in line[:11].split("-") ]
-                words = line[11:].split()
+                atoms, words = _frcmod_atoms_words(line, 11)
                 propers += "-".join(atoms) + "\t" + str(float(words[1]) / int(words[0])) + "\t" + words[2] + "\t" + str(abs(int(float(words[3])))) + "\t" + str(reset) + "\n"
                 if int(float(words[3])) < 0:
                     reset = 0
                 else:
                     reset = 1
             elif flag[:4] == "IMPR":
-                atoms = [ word.strip() for word in line[:11].split("-") ]
-                words = line[11:].split()
+                atoms, words = _frcmod_atoms_words(line, 11)
[…]
