+++ Xponge/BUILD.py
@@ -13,78 +13,87 @@ def _analyze_connectivity(cls):
                 index_next.update(index_temp)
             index_dict = index_next
 
+def _get_frc_all(frc, cls):
+    top = frc.topology_like
+    top_matrix = frc.topology_matrix
+    frc_all = []
+    for atom0 in cls.atoms:
+        backups = {i: [] for i in range(len(top))}
+        backups[0].append([atom0])
+        for i, d in enumerate(top):
+            if i == 0:
+                continue
+            for atom1 in atom0.linked_atoms[d]:
+                for backup in backups[i - 1]:
+                    good_backup = True
+                    for j, atomj in enumerate(backup):
+                        if atomj == atom1 or abs(top_matrix[j][i]) <= 1 \
+                                or atom1 not in atomj.linked_atoms[abs(top_matrix[j][i])]:
+                            good_backup = False
+                            break
+                        if top_matrix[j][i] <= -1:
+                            for d2 in range(2, d):
+                                if atom1 in atomj.linked_atoms[d2]:
+                                    good_backup = False
+                                    break
+                    if good_backup:
+                        backups[i].append([*backup, atom1])
+        frc_all.extend(backups[len(top) - 1])
+    return frc_all
+
+def _get_frc_all_final(frc, frc_all):
+    frc_all_final = []
+    frc_keys = {}
+    for frc_one in frc_all:
+        frc_one_name = "".join([str(hash(atom)) for atom in frc_one])
+        if frc_one_name in frc_keys.keys():
+            frc_keys[frc_one_name].append(frc_one)
+        else:
+            temp_list = [frc_one]
+            frc_all_final.append(temp_list)
+            for atom_permutation in frc.Same_Force(frc_one):
+                frc_one_name = "".join([str(hash(atom)) for atom in atom_permutation])
+                frc_keys[frc_one_name] = temp_list
+    return frc_all_final
+
+def _find_the_force(frc, frc_all_final, cls):
+    for frc_ones in frc_all_final:
+        finded = {}
+        # 先直接找
+        for frc_one in frc_ones:
+            tofindname = "-".join([atom.type.name for atom in frc_one])
+            if tofindname in frc.types.keys():
+                finded[tofindname] = [frc.types[tofindname], frc_one]
+                break
+        # 没找到再找通用的
+        if not finded:
+            leastfindedX = 999
+            for frc_one in frc_ones:
+                tofind = [[atom.type.name, "X"] for atom in frc_one]
+                for p in product(*tofind):
+                    pcountx = p.count("X")
+                    if pcountx > leastfindedX:
+                        continue
+                    tofindname = "-".join(p)
+                    if tofindname in frc.types.keys():
+                        finded = {tofindname: [frc.types[tofindname], frc_one]}
+                        leastfindedX = pcountx
+                        break
+
+        assert (not frc.compulsory or len(finded) == 1), "None of %s type found for %s" % (
+            frc.name, "-".join([atom.type.name for atom in frc_one]))
+
+        if finded:
+            for finded_type, finded_atoms in finded.values():
+                cls.Add_Bonded_Force(frc.entity(finded_atoms, finded_type))
 
 def _build_bfrc(cls):
     _analyze_connectivity(cls)
     for frc in GlobalSetting.BondedForces:
-        top = frc.topology_like
-        top_matrix = frc.topology_matrix
-        frc_all = []
-        for atom0 in cls.atoms:
-            backups = {i: [] for i in range(len(top))}
-            backups[0].append([atom0])
-            for i, d in enumerate(top):
-                if i == 0:
-                    continue
-                for atom1 in atom0.linked_atoms[d]:
[…]
