+++ docs/source/Xponge.forcefield.AMBER.rst
@@ -1,0 +1,85 @@
+Xponge.forcefield.AMBER package
+===============================
+
+Submodules
+----------
+
+Xponge.forcefield.AMBER.ff14SB module
+-------------------------------------
+
+.. automodule:: Xponge.forcefield.AMBER.ff14SB
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.AMBER.ff19SB module
+-------------------------------------
+
+.. automodule:: Xponge.forcefield.AMBER.ff19SB
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.AMBER.gaff module
+-----------------------------------
+
+.. automodule:: Xponge.forcefield.AMBER.gaff
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.AMBER.lipid14 module
+--------------------------------------
+
+.. automodule:: Xponge.forcefield.AMBER.lipid14
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.AMBER.lipid17 module
+--------------------------------------
+
+.. automodule:: Xponge.forcefield.AMBER.lipid17
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.AMBER.opc module
+----------------------------------
+
+.. automodule:: Xponge.forcefield.AMBER.opc
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.AMBER.spce module
+-----------------------------------
+
+.. automodule:: Xponge.forcefield.AMBER.spce
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.AMBER.tip3p module
+------------------------------------
+
+.. automodule:: Xponge.forcefield.AMBER.tip3p
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.AMBER.tip4pew module
+--------------------------------------
+
+.. automodule:: Xponge.forcefield.AMBER.tip4pew
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: Xponge.forcefield.AMBER
+   :members:
+   :undoc-members:
+   :show-inheritance:
