+++ docs/source/Xponge.forcefield.rst
@@ -1,0 +1,21 @@
+Xponge.forcefield package
+=========================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   Xponge.forcefield.AMBER
+   Xponge.forcefield.BASE
+   Xponge.forcefield.CHARMM27
+   Xponge.forcefield.SPECIAL
+
+Module contents
+---------------
+
+.. automodule:: Xponge.forcefield
+   :members:
+   :undoc-members:
+   :show-inheritance:
