+++ docs/source/Xponge.forcefield.BASE.rst
@@ -1,0 +1,133 @@
+Xponge.forcefield.BASE package
+==============================
+
+Submodules
+----------
+
+Xponge.forcefield.BASE.ACMAP module
+-----------------------------------
+
+.. automodule:: Xponge.forcefield.BASE.ACMAP
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.BASE.ANGLE module
+-----------------------------------
+
+.. automodule:: Xponge.forcefield.BASE.ANGLE
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.BASE.BOND module
+----------------------------------
+
+.. automodule:: Xponge.forcefield.BASE.BOND
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.BASE.CHARGE module
+------------------------------------
+
+.. automodule:: Xponge.forcefield.BASE.CHARGE
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.BASE.DIHEDRAL module
+--------------------------------------
+
+.. automodule:: Xponge.forcefield.BASE.DIHEDRAL
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.BASE.EXCLUDE module
+-------------------------------------
+
+.. automodule:: Xponge.forcefield.BASE.EXCLUDE
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.BASE.IMPROPER module
+--------------------------------------
+
+.. automodule:: Xponge.forcefield.BASE.IMPROPER
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.BASE.LJ module
+--------------------------------
+
+.. automodule:: Xponge.forcefield.BASE.LJ
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.BASE.MASS module
+----------------------------------
+
+.. automodule:: Xponge.forcefield.BASE.MASS
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.BASE.NB14 module
+----------------------------------
+
+.. automodule:: Xponge.forcefield.BASE.NB14
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.BASE.NB14\_EXTRA module
+-----------------------------------------
+
+.. automodule:: Xponge.forcefield.BASE.NB14_EXTRA
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.BASE.RCMAP module
+-----------------------------------
+
+.. automodule:: Xponge.forcefield.BASE.RCMAP
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.BASE.SOFT\_BOND module
+----------------------------------------
+
+.. automodule:: Xponge.forcefield.BASE.SOFT_BOND
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.BASE.UREY\_BRADLEY module
+-------------------------------------------
+
+.. automodule:: Xponge.forcefield.BASE.UREY_BRADLEY
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.BASE.VIRTUAL\_ATOM module
+-------------------------------------------
+
+.. automodule:: Xponge.forcefield.BASE.VIRTUAL_ATOM
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: Xponge.forcefield.BASE
+   :members:
+   :undoc-members:
+   :show-inheritance:
