+++ docs/source/Xponge.rst
@@ -1,0 +1,56 @@
+Xponge package
+==============
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   Xponge.analysis
+   Xponge.assign
+   Xponge.forcefield
+   Xponge.tools
+
+Submodules
+----------
+
+Xponge.BUILD module
+-------------------
+
+.. automodule:: Xponge.BUILD
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.IMPOSE module
+--------------------
+
+.. automodule:: Xponge.IMPOSE
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.LOAD module
+------------------
+
+.. automodule:: Xponge.LOAD
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.PROCESS module
+---------------------
+
+.. automodule:: Xponge.PROCESS
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: Xponge
+   :members:
+   :undoc-members:
+   :show-inheritance:
