+++ docs/source/modules.rst
@@ -1,0 +1,7 @@
+Xponge
+======
+
+.. toctree::
+   :maxdepth: 4
+
+   Xponge

[…]
