+++ Xponge/forcefield/AMBER/ff19SB.py
@@ -11,7 +11,6 @@ DIHEDRAL.ProperType.New_From_String(propers)
 DIHEDRAL.ImproperType.New_From_String(impropers)
 LJ.LJType.New_From_String(LJs)
 
-
 atoms, bonds, angles, propers, impropers, LJs, cmap = LOAD.frcmod(os.path.join(AMBER_DATA_DIR, "ff19SB.frcmod"))
 
 AtomType.New_From_String(atoms)
@@ -22,6 +21,7 @@ DIHEDRAL.ImproperType.New_From_String(impropers)
 LJ.LJType.New_From_String(LJs)
 
 from ..BASE import RCMAP
+
 RCMAP.CMAP.Residue_Map.update(cmap)
 
 RCMAP.CMAP.New_From_String(r"""
@@ -29,14 +29,13 @@ name
 C-N-XC-C-N
 """)
 
-
 ff19SB = LOAD.mol2(os.path.join(AMBER_DATA_DIR, "ff19SB.mol2"))
 ResidueType.types["HIS"] = ResidueType.types["HIE"]
 ResidueType.types["NHIS"] = ResidueType.types["NHIE"]
 ResidueType.types["CHIS"] = ResidueType.types["CHIE"]
-sys.modules['__main__'].__dict__["HIS"] = ResidueType.types["HIS"] 
-sys.modules['__main__'].__dict__["NHIS"] = ResidueType.types["NHIS"] 
-sys.modules['__main__'].__dict__["CHIS"] = ResidueType.types["CHIS"] 
+sys.modules['__main__'].__dict__["HIS"] = ResidueType.types["HIS"]
+sys.modules['__main__'].__dict__["NHIS"] = ResidueType.types["NHIS"]
+sys.modules['__main__'].__dict__["CHIS"] = ResidueType.types["CHIS"]
 
 residues = "ALA ARG ASN ASP CYS CYX GLN GLU GLY HID HIE HIP ILE LEU LYS MET PHE PRO SER THR TRP TYR VAL HIS".split()
 
@@ -45,43 +44,41 @@ for res in residues:
     ResidueType.types[res].head_length = 1.3
     ResidueType.types[res].tail_next = "CA"
     ResidueType.types[res].tail_length = 1.3
-    ResidueType.types[res].head_link_conditions.append({"atoms":["CA", "N"], "parameter": 120/180 * np.pi})
-    ResidueType.types[res].head_link_conditions.append({"atoms":["H", "CA", "N"], "parameter": -np.pi})
-    ResidueType.types[res].tail_link_conditions.append({"atoms":["CA", "C"], "parameter": 120/180 * np.pi})
-    ResidueType.types[res].tail_link_conditions.append({"atoms":["O", "CA", "C"], "parameter": -np.pi})    
-
+    ResidueType.types[res].head_link_conditions.append({"atoms": ["CA", "N"], "parameter": 120 / 180 * np.pi})
+    ResidueType.types[res].head_link_conditions.append({"atoms": ["H", "CA", "N"], "parameter": -np.pi})
+    ResidueType.types[res].tail_link_conditions.append({"atoms": ["CA", "C"], "parameter": 120 / 180 * np.pi})
+    ResidueType.types[res].tail_link_conditions.append({"atoms": ["O", "CA", "C"], "parameter": -np.pi})
 
     ResidueType.types["N" + res].tail_next = "CA"
     ResidueType.types["N" + res].tail_length = 1.3
-    ResidueType.types["N" + res].tail_link_conditions.append({"atoms":["CA", "C"], "parameter": 120/180 * np.pi})
-    ResidueType.types["N" + res].tail_link_conditions.append({"atoms":["O", "CA", "C"], "parameter": -np.pi})   
-    
[…]
