+++ Xponge/assign/__init__.py
@@ -29,7 +29,7 @@ class _RING():
             for atom in ring.atoms:
                 assign.Add_Atom_Marker(atom, "RG")
                 assign.Add_Atom_Marker(atom, "RG%d" % len(ring.atoms))
-        
+
     @staticmethod
     def Check_Rings_Type(assign, rings):
         for ring in rings:
@@ -48,7 +48,7 @@ class _RING():
                             assign.Add_Atom_Marker(atom, "AR2")
                     else:
                         assign.Add_Atom_Marker(atom, "AR4")
-        
+
     @staticmethod
     def Get_Rings(assign):
         current_path = []
@@ -89,7 +89,7 @@ class _RING():
                             father_atom = current_path_father.pop()
                             current_path_sons[father_atom] -= 1
         return have_found_rings
-        
+
     def __repr__(self):
         return self.tohash
 
@@ -111,12 +111,13 @@ class _RING():
     def get_3_neighbors(self):
         for i in range(len(self.atoms)):
             yield self.atoms[i - 2], self.atoms[i - 1], self.atoms[i]
-    
+
     def check_pure_aromatic(self, assign):
         if len(self.atoms) == 6:
             self.is_pure_aromatic_ring = True
             for atom in self.atoms:
-                if not assign.Atom_Judge(atom, "C3") and not assign.Atom_Judge(atom, "N2") and not assign.Atom_Judge(atom, "N3"):
+                if not assign.Atom_Judge(atom, "C3") and not assign.Atom_Judge(atom, "N2") and not assign.Atom_Judge(
+                        atom, "N3"):
                     self.is_pure_aromatic_ring = False
                     break
                 if assign.Atom_Judge(atom, "N3"):
@@ -134,7 +135,7 @@ class _RING():
                     break
         else:
             self.is_pure_aromatic_ring = False
-    
+
     def check_pure_aliphatic_and_planar(self, assign):
         self.is_pure_aliphatic_ring = True
         self.is_planar_ring = True
@@ -150,6 +151,7 @@ class _RING():
                     and not assign.Atom_Judge(atom, "S2") and not assign.Atom_Judge(atom, "P2")
                     and not assign.Atom_Judge(atom, "P3")):
                 self.is_planar_ring = False
+
     def check_out_plane_double_bond(self, assign):
         self.out_plane_double_bond = False
         for atom in self.atoms:
@@ -166,7 +168,7 @@ class Assign():
     XD = set("SP")
     XE = set(["N", "O", "F", "Cl", "Br", "S", "I"])
 
-    def __init__(self, name = "ASN"):
+    def __init__(self, name="ASN"):
         self.name = name
         self.atom_numbers = 0
         self.atoms = []
@@ -199,7 +201,7 @@ class Assign():
                 break
         return judge
 
-    def Add_Atom(self, element, x, y, z, name="", charge = 0.0):
+    def Add_Atom(self, element, x, y, z, name="", charge=0.0):
         if "." in element:
             element, element_detail = element.split(".")
             element_detail = "." + element_detail
@@ -221,7 +223,7 @@ class Assign():
             self.charge = np.array([charge])
         else:
             self.charge = np.hstack((self.charge, np.array([charge])))
- 
+
     def Add_Atom_Marker(self, atom, marker):
[…]
