+++ Xponge/forcefield/SPECIAL/FEP.py
@@ -12,6 +12,7 @@ ZERO_LJ_ATOM-ZERO_LJ_ATOM  0  0
 AtomType.Add_Property({"LJtypeB": str})
 AtomType.Add_Property({"subsys": int})
 
+
 def _find_common_forces(forcetype, Aforces, Bforces, molB2molA):
     toret = []
     temp_map = {}
@@ -87,7 +88,7 @@ def nb14_extra_merge_rule(molR, molA, molB, forcetype, Rforces, Bforces, _lambda
                 nb14_ee_factor = fB.kee * fB.atoms[0].charge * \
                                  fB.atoms[1].charge
             kee = nb14_ee_factor / temp_charge0 / temp_charge1
-            
+
             fR.kee = fR.kee * (1 - _lambda) + kee * _lambda
             fR.A = fR.A * (1 - _lambda) + fB.A * _lambda
             fR.B = fR.B * (1 - _lambda) + fB.B * _lambda
@@ -136,7 +137,7 @@ def angle_merge_rule(molR, molA, molB, forcetype, Rforces, Bforces, _lambda, mol
             fR.b = fB.b
             molR.Add_Bonded_Force(fR)
         elif abs(fR.b - fB.b) < 1e-5:
-            fR.k = fR.k * (1 - _lambda) + fB.k * _lambda 
+            fR.k = fR.k * (1 - _lambda) + fB.k * _lambda
         else:
             fR2 = ANGLE.AngleType.entity(list(map(lambda x: molB2molR[x], fB.atoms)), fB.type, fB.name)
             fR2.k = fB.k * _lambda
@@ -224,17 +225,18 @@ def Save_Hard_Core_LJ():
     Molecule.Del_Save_SPONGE_Input("LJ_soft_core")
     Molecule.Del_Save_SPONGE_Input("subsys_division")
 
+
 def Save_Soft_Core_LJ():
     Molecule.Del_Save_SPONGE_Input("LJ")
 
     @Molecule.Set_Save_SPONGE_Input("subsys_division")
     def write_subsys_division(self):
-        towrite = "%d\n"%len(self.atoms)
+        towrite = "%d\n" % len(self.atoms)
         for atom in self.atoms:
             if getattr(atom, "subsys", None) is None:
-                towrite += "%d\n"%0
+                towrite += "%d\n" % 0
             else:
-                towrite += "%d\n"%atom.subsys
+                towrite += "%d\n" % atom.subsys
         return towrite
 
     @Molecule.Set_Save_SPONGE_Input("LJ_soft_core")
@@ -252,7 +254,7 @@ def Save_Soft_Core_LJ():
             if atom.LJtypeB not in LJtypemapB.keys():
                 LJtypemapB[atom.LJtypeB] = len(LJtypesB)
                 LJtypesB.append(atom.LJtypeB)
-        
+
         As, Bs = LJ.find_AB_LJ(LJtypes)
         AsB, BsB = LJ.find_AB_LJ(LJtypesB)
 
@@ -265,7 +267,7 @@ def Save_Soft_Core_LJ():
         same_typeB = LJ.judge_same_type(LJtypesB, checks)
         real_LJtypesB = LJ.get_real_LJ(LJtypesB, same_typeB)
         real_AsB, real_BsB = LJ.find_AB_LJ(real_LJtypesB, False)
-        
+
         towrite = "%d %d %d\n\n" % (len(self.atoms), len(real_LJtypes), len(real_LJtypesB))
         count = 0
         for i in range(len(real_LJtypes)):
@@ -349,6 +351,7 @@ def Get_Free_Molecule(molA, perturbing_residues, intra_FEP=False):
 
     return molB
 
+
[…]
