+++ Xponge/LOAD.py
@@ -416,341 +416,6 @@ def parmdat(filename):
 sys.modules['__main__'].__dict__["loadparmdat"] = parmdat
 
 
-def _read_parm7_someflag(length, word_process, offset, f):
-    tempvar = []
-    line = f.readline()
-    while line.startswith("%"):
-        line = f.readline()
-    while 1:
-        i = 0
-        while i + offset < len(line):
-            word = line[i:i + offset].strip()
-            if word:
-                tempvar.append(word_process(word))
-            else:
-                break
-            i += offset
-        if len(tempvar) == length:
-            break
-        line = f.readline()
-    return tempvar
-
-
-def _read_parm7_build(mol, BONDTYPE, bonds, parameter_head, parameter_lists, add_connectivity=False):
-    bond_names = {}
-    newBONDs = ""
-    newBONDs += parameter_head
-    for i in bonds.keys():
-        bond_names[i] = "-".join([mol.atoms[atype].type.name for atype in bonds[i][0]])
-        newBONDs += bond_names[i]
-        for fmt, parm_list in parameter_lists:
-            newBONDs += " " + fmt % parm_list[i]
-        newBONDs += "\n"
-    BONDTYPE.New_From_String(newBONDs)
-
-    for typei, bondi in bonds.items():
-        for atom_indexes in bondi:
-            atoms = [mol.atoms[i] for i in atom_indexes]
-            residues = set([atom.residue for atom in atoms])
-            if len(residues) == 1:
-                residue = atoms[0].residue
-                resatoms = [residue.type._name2atom[atom.name] for atom in atoms]
-                residue.type.Add_Bonded_Force(
-                    BONDTYPE.entity(resatoms, BONDTYPE.types_different_name[bond_names[typei]]))
-                residue.Add_Bonded_Force(BONDTYPE.entity(atoms, BONDTYPE.types_different_name[bond_names[typei]]))
-                if add_connectivity:
-                    residue.Add_Connectivity(atoms[0], atoms[-1])
-            else:
-                raise NotImplementedError
-
-
-def _read_parm7_read(filename):
-    bonds = {}
-    angles = {}
-    propers = {}
-    nb14s = {}
-    impropers = {}
-    with open(filename) as f:
-        for line in f:
-            if line.startswith("%FLAG TITLE"):
-                line = f.readline()
-                while line.startswith("%"):
-                    line = f.readline()
-                try:
-                    name = line.split()[0].strip()
-                except IndexError:
-                    from random import randint
-                    name = "Molecule%d" % randint(0, 10086)
-                mol = Molecule(name=name)
-            elif line.startswith("%FLAG POINTERS"):
-                line = f.readline()
-                while line.startswith("%"):
-                    line = f.readline()
-                atom_numbers = int(line[0:8])
-                atom_type_numbers = int(line[8:16])
-                bond_numbers_H = int(line[16:24])
-                bond_numbers_N = int(line[24:32])
-                bond_numbers = bond_numbers_H + bond_numbers_N
-                angle_numbers_H = int(line[32:40])
-                angle_numbers_N = int(line[40:48])
-                angle_numbers = angle_numbers_H + angle_numbers_N
-                dihedral_numbers_H = int(line[48:56])
-                dihedral_numbers_N = int(line[56:64])
[…]
