### 2022-07-18 5fbd781
clean build.py
