+++ Xponge/helper.py
@@ -1,3 +1,711 @@
-from . import GlobalSetting
+import sys
+import time
+import numpy as np
+from collections import OrderedDict
+from itertools import product
 
 
+class _GlobalSetting():
+    # 最远的成键距离，用于拓扑分析时最远分析多远
+    farthest_bonded_force = 0
+    # 所有的成键类型力的Type
+    BondedForces = []
+    BondedForcesMap = {}
+    # 所有虚拟原子的Type和对应的依赖的其他原子的数量
+    VirtualAtomTypes = {}
+    # 单位换算
+    UnitMapping = {"distance": {"nm": 1e-9, "A": 1e-10},
+                   "energy": {"kcal/mol": 4.184, "eV": 96.4853, "kJ/mol": 1},
+                   "charge": {"e": 1, "SPONGE": 1.0 / 18.2223},
+                   "angle": {"degree": np.pi, "rad": 180}
+                   }
+    PDBResidueNameMap = {"head": {}, "tail": {}, "save": {}}
+    HISMap = {"DeltaH": "", "EpsilonH": "", "HIS": {}}
+    nocenter = False
+    boxspace = 3  # space between the molecule and the box border
+
+    def Add_PDB_Residue_Name_Mapping(self, place, pdb_name, real_name):
+        assert place in ("head", "tail")
+        self.PDBResidueNameMap[place][pdb_name] = real_name
+        self.PDBResidueNameMap["save"][real_name] = pdb_name
+
+    def Set_Invisible_Bonded_Forces(self, types):
+        for typename in types:
+            self.BondedForces.remove(self.BondedForcesMap[typename])
+
+    def Set_Visible_Bonded_Forces(self, types):
+        self.BondedForces = []
+        for typename in types:
+            self.BondedForces.append(self.BondedForcesMap[typename])
+
+    @staticmethod
+    def Set_Unit_Transfer_Function(sometype):
+        def wrapper(func):
+            sometype._unit_transfer = func
+
+        return wrapper
+
+    @staticmethod
+    def Add_Unit_Transfer_Function(sometype):
+        func0 = sometype._unit_transfer
+
+        def wrapper(func):
+            def temp(self):
+                func0(self)
+                func(self)
+
+            sometype._unit_transfer = temp
+
+        return wrapper
+
+
+GlobalSetting = _GlobalSetting()
+
+
+class Type:
+    name = None
+    parameters = {"name": str}
+    types = {}
+    types_different_name = {}
+
+    @classmethod
+    def Add_Property(cls, parm_fmt, parm_default=None):
+        cls.parameters.update(parm_fmt)
+        if parm_default is None:
+            parm_default = {}
+        for _type in cls.types.values():
+            _type.contents.update({key: parm_default.get(key, None) for key in parm_fmt.keys()})
+
+    @classmethod
+    def Set_Property_Unit(cls, prop, unit_type, base_unit):
+        assert prop in cls.parameters.keys(), "Unknown property '%s' for type '%s'" % (prop, cls.name)
+        temp_unit_lists = unit_type.split("·")
+        temp_unit_power = []
+
+        for i in range(len(temp_unit_lists)):
+            unit = temp_unit_lists[i].split("^")
+            if len(unit) == 2:
+                temp_unit_lists[i] = unit[0]
+                temp_unit_power.append(int(unit[1]))
+            else:
+                temp_unit_power.append(1)
+            assert unit[0] in GlobalSetting.UnitMapping.keys(), "Unknown unit type '%s'" % unit_type
+
+        temp_unit_lists_units = [list(GlobalSetting.UnitMapping[unit].keys()) for unit in temp_unit_lists]
+        alls = {}
+        for unit_combination in product(*temp_unit_lists_units):
+            unit = []
+            value = 1
[…]
