+++ Xponge/build.py
@@ -1,8 +1,10 @@
 """
 this **module** is used to build and save
 """
-from . import *
-import sys
+import os
+from itertools import product
+from . import assign
+from .helper import ResidueType, Molecule, Residue, ResidueLink, GlobalSetting
 
 
 def _analyze_connectivity(cls):
@@ -23,6 +25,31 @@ def _analyze_connectivity(cls):
             index_dict = index_next
 
 
+def _check_backup(backups, atom1, top_matrix, i, d):
+    """
+
+    :param backups:
+    :param atom1:
+    :param top_matrix:
+    :param d:
+    :return:
+    """
+    for backup in backups[i - 1]:
+        good_backup = True
+        for j, atomj in enumerate(backup):
+            if atomj == atom1 or abs(top_matrix[j][i]) <= 1 \
+                    or atom1 not in atomj.linked_atoms[abs(top_matrix[j][i])]:
+                good_backup = False
+                break
+            if top_matrix[j][i] <= -1:
+                for d2 in range(2, d):
+                    if atom1 in atomj.linked_atoms[d2]:
+                        good_backup = False
+                        break
+        if good_backup:
+            backups[i].append([*backup, atom1])
+
+
 def _get_frc_all(frc, cls):
     """
 
@@ -40,20 +67,8 @@ def _get_frc_all(frc, cls):
             if i == 0:
                 continue
             for atom1 in atom0.linked_atoms[d]:
-                for backup in backups[i - 1]:
-                    good_backup = True
-                    for j, atomj in enumerate(backup):
-                        if atomj == atom1 or abs(top_matrix[j][i]) <= 1 \
-                                or atom1 not in atomj.linked_atoms[abs(top_matrix[j][i])]:
-                            good_backup = False
-                            break
-                        if top_matrix[j][i] <= -1:
-                            for d2 in range(2, d):
-                                if atom1 in atomj.linked_atoms[d2]:
-                                    good_backup = False
-                                    break
-                    if good_backup:
-                        backups[i].append([*backup, atom1])
+                _check_backup(backups, atom1, top_matrix, i, d)
+
         frc_all.extend(backups[len(top) - 1])
     return frc_all
 
@@ -98,17 +113,17 @@ def _find_the_force(frc, frc_all_final, cls):
                 break
         # 没找到再找通用的
         if not finded:
-            leastfindedX = 999
+            leastfinded_x = 999
             for frc_one in frc_ones:
                 tofind = [[atom.type.name, "X"] for atom in frc_one]
                 for p in product(*tofind):
                     pcountx = p.count("X")
-                    if pcountx > leastfindedX:
+                    if pcountx > leastfinded_x:
                         continue
                     tofindname = "-".join(p)
                     if tofindname in frc.types.keys():
                         finded = {tofindname: [frc.types[tofindname], frc_one]}
-                        leastfindedX = pcountx
+                        leastfinded_x = pcountx
                         break
 
         assert (not frc.compulsory or len(finded) == 1), "None of %s type found for %s" % (
@@ -119,7 +134,7 @@ def _find_the_force(frc, frc_all_final, cls):
                 cls.Add_Bonded_Force(frc.entity(finded_atoms, finded_type))
 
 
-def _build_bfrc(cls):
+def _build_residue_type(cls):
     """
 
     :param cls:
@@ -132,14 +147,14 @@ def _build_bfrc(cls):
         _find_the_force(frc, frc_all_final, cls)
 
 
[…]

[…]
