### 2022-07-19 29b8b0a
clean load.py
