+++ Xponge/forcefield/AMBER/opc.py
@@ -34,9 +34,9 @@ name   atom0    atom1   atom2   k1         k2
 EP     -3       -2      -1      0.1477206  0.1477206
 """)
 
-TIP3P = LOAD.mol2(os.path.join(AMBER_DATA_DIR, "opc.mol2"))
+TIP3P = load_mol2(os.path.join(AMBER_DATA_DIR, "opc.mol2"))
 
-atoms, bonds, angles, propers, impropers, LJs, cmap = LOAD.frcmod(
+atoms, bonds, angles, propers, impropers, LJs, cmap = load_frcmod(
     os.path.join(AMBER_DATA_DIR, "ions1lm_126_tip4pew.frcmod"))
 AtomType.New_From_String(atoms)
 BOND.BondType.New_From_String(bonds)
@@ -45,7 +45,7 @@ DIHEDRAL.ProperType.New_From_String(propers)
 DIHEDRAL.ImproperType.New_From_String(impropers)
 LJ.LJType.New_From_String(LJs)
 
-atoms, bonds, angles, propers, impropers, LJs, cmap = LOAD.frcmod(os.path.join(AMBER_DATA_DIR, "ionsjc_tip4pew.frcmod"))
+atoms, bonds, angles, propers, impropers, LJs, cmap = load_frcmod(os.path.join(AMBER_DATA_DIR, "ionsjc_tip4pew.frcmod"))
 AtomType.New_From_String(atoms)
 BOND.BondType.New_From_String(bonds)
 ANGLE.AngleType.New_From_String(angles)
@@ -53,7 +53,7 @@ DIHEDRAL.ProperType.New_From_String(propers)
 DIHEDRAL.ImproperType.New_From_String(impropers)
 LJ.LJType.New_From_String(LJs)
 
-atoms, bonds, angles, propers, impropers, LJs, cmap = LOAD.frcmod(
+atoms, bonds, angles, propers, impropers, LJs, cmap = load_frcmod(
     os.path.join(AMBER_DATA_DIR, "ions234lm_126_tip4pew.frcmod"))
 AtomType.New_From_String(atoms)
 BOND.BondType.New_From_String(bonds)
@@ -62,7 +62,7 @@ DIHEDRAL.ProperType.New_From_String(propers)
 DIHEDRAL.ImproperType.New_From_String(impropers)
 LJ.LJType.New_From_String(LJs)
 
-LOAD.mol2(os.path.join(AMBER_DATA_DIR, "atomic_ions.mol2"))
+load_mol2(os.path.join(AMBER_DATA_DIR, "atomic_ions.mol2"))
 
 print("""Reference for opc:
 1. Water:
