+++ Xponge/forcefield/AMBER/tip3p.py
@@ -27,9 +27,9 @@ OW-OW   0.152               1.7683
 HW-HW   0                   0
 """)
 
-TIP3P = LOAD.mol2(os.path.join(AMBER_DATA_DIR, "tip3p.mol2"))
+TIP3P = load_mol2(os.path.join(AMBER_DATA_DIR, "tip3p.mol2"))
 
-atoms, bonds, angles, propers, impropers, LJs, cmap = LOAD.frcmod(
+atoms, bonds, angles, propers, impropers, LJs, cmap = load_frcmod(
     os.path.join(AMBER_DATA_DIR, "ions1lm_126_tip3p.frcmod"))
 AtomType.New_From_String(atoms)
 BOND.BondType.New_From_String(bonds)
@@ -38,7 +38,7 @@ DIHEDRAL.ProperType.New_From_String(propers)
 DIHEDRAL.ImproperType.New_From_String(impropers)
 LJ.LJType.New_From_String(LJs)
 
-atoms, bonds, angles, propers, impropers, LJs, cmap = LOAD.frcmod(os.path.join(AMBER_DATA_DIR, "ionsjc_tip3p.frcmod"))
+atoms, bonds, angles, propers, impropers, LJs, cmap = load_frcmod(os.path.join(AMBER_DATA_DIR, "ionsjc_tip3p.frcmod"))
 AtomType.New_From_String(atoms)
 BOND.BondType.New_From_String(bonds)
 ANGLE.AngleType.New_From_String(angles)
@@ -46,7 +46,7 @@ DIHEDRAL.ProperType.New_From_String(propers)
 DIHEDRAL.ImproperType.New_From_String(impropers)
 LJ.LJType.New_From_String(LJs)
 
-atoms, bonds, angles, propers, impropers, LJs, cmap = LOAD.frcmod(
+atoms, bonds, angles, propers, impropers, LJs, cmap = load_frcmod(
     os.path.join(AMBER_DATA_DIR, "ions234lm_126_tip3p.frcmod"))
 AtomType.New_From_String(atoms)
 BOND.BondType.New_From_String(bonds)
@@ -55,7 +55,7 @@ DIHEDRAL.ProperType.New_From_String(propers)
 DIHEDRAL.ImproperType.New_From_String(impropers)
 LJ.LJType.New_From_String(LJs)
 
-LOAD.mol2(os.path.join(AMBER_DATA_DIR, "atomic_ions.mol2"))
+load_mol2(os.path.join(AMBER_DATA_DIR, "atomic_ions.mol2"))
 
 print("""Reference for tip3p:
 1. Water:
