+++ Xponge/build.py
@@ -1,5 +1,5 @@
 """
-this **module** is used to build and save
+This **module** is used to build and save
 """
 import os
 from itertools import product

[…]
