+++ Xponge/helper.py
@@ -603,6 +603,40 @@ class Molecule:
     all = {}
     save_functions = {}
 
+    @staticmethod
+    def _set_friends_in_different_residue(molecule, atom1, atom2):
+        if molecule.atom_index[atom1.residue.atoms[0]] < molecule.atom_index[atom1.residue.atoms[1]]:
+            res_index = molecule.atom_index[atom1.residue.atoms[-1]]
+            atom1_friends = list(range(res_index + 1))
+            atom2_friends = list(range(res_index + 1, len(molecule.atoms)))
+        else:
+            res_index = molecule.atom_index[atom2.residue.atoms[-1]]
+            atom2_friends = list(range(res_index + 1))
+            atom1_friends = list(range(res_index + 1, len(molecule.atoms)))
+        return atom1_friends, atom2_friends
+
+    @staticmethod
+    def _get_head_and_tail(head, tail, molecule, typeatom1, typeatom2, restype, toset, atom1_friends, _resatom):
+        assert typeatom2 in restype.connectivity[typeatom1]
+        index_dict = {}.fromkeys(restype.connectivity[typeatom1], typeatom1)
+        if typeatom2 in index_dict.keys():
+            index_dict.pop(typeatom2)
+
+        while index_dict:
+            index_next = {}
+            for atom0, from_atom in index_dict.items():
+                if atom0.name == restype.head:
+                    head = toset
+                elif atom0.name == restype.tail:
+                    tail = toset
+                atom1_friends.append(molecule.atom_index[_resatom(atom0)])
+                index_temp = {}.fromkeys(restype.connectivity[atom0], atom0)
+                index_temp.pop(from_atom)
+                if typeatom2 in index_temp.keys():
+                    index_temp.pop(typeatom2)
+                index_next.update(index_temp)
+            index_dict = index_next
+        return head, tail
     @classmethod
     def Set_Save_SPONGE_Input(cls, keyname):
         def wrapper(func):
@@ -693,10 +727,420 @@ class Molecule:
                 atom.copied.pop(forcopy)
         return new_molecule
 
+    def get_atom_coordinates(self):
+        self.atoms = []
+        for res in self.residues:
+            self.atoms.extend(res.atoms)
+
+        self.atom_index = {self.atoms[i]: i for i in range(len(self.atoms))}
+        return np.array([[atom.x, atom.y, atom.z] for atom in self.atoms])
+
+
+    def divide_into_two_parts(self, atom1, atom2):
+        if atom1.residue != atom2.residue:
+            atom1_friends, atom2_friends = self._set_friends_in_different_residue(self, atom1, atom2)
+        else:
+            link_front = 0
+            atom1_friends = []
+            atom2_friends = []
+            head = 0
+            tail = 0
+
+            restype = atom1.residue.type
+
+            _restype_atom = lambda atom: restype._name2atom[atom.name]
+            _resatom = lambda atom: atom1.residue._name2atom[atom.name]
+
+            typeatom1 = _restype_atom(atom1)
+            typeatom2 = _restype_atom(atom2)
+
+            head, tail = self._get_head_and_tail(head, tail, self, typeatom1, typeatom2, restype, 1, atom1_friends,
+                                            _resatom)
+            head, tail = self._get_head_and_tail(head, tail, self, typeatom2, typeatom1, restype, 2, atom2_friends,
+                                            _resatom)
+
[…]
