+++ Xponge/process.py
@@ -1,0 +1,269 @@
+"""
+This **module** is used to process topology and conformations
+"""
+import numpy as np
+from .helper import get_rotate_matrix, ResidueType, Molecule, Residue
+
+
+def impose_bond(molecule, atom1, atom2, length):
+    """
+
+    :param molecule:
+    :param atom1:
+    :param atom2:
+    :param length:
+    :return:
+    """
+    crd = molecule.get_atom_coordinates()
+    _, atom2_friends = molecule.divide_into_two_parts(atom1, atom2)
+    r0 = crd[molecule.atom_index[atom2]] - crd[molecule.atom_index[atom1]]
+    l0 = np.linalg.norm(r0)
+    if l0 == 0:
+        crd[molecule.atom_index[atom2]] += (1 / 3) ** (0.5)
+        r0 = crd[molecule.atom_index[atom2]] - crd[molecule.atom_index[atom1]]
+        l0 = np.linalg.norm(r0)
+    dr = (length / l0 - 1) * r0
+    crd[atom2_friends] += dr
+    for atom in molecule.atoms:
+        i = molecule.atom_index[atom]
+        atom.x = crd[i][0]
+        atom.y = crd[i][1]
+        atom.z = crd[i][2]
+
+
+def impose_angle(molecule, atom1, atom2, atom3, angle):
+    """
+
+    :param molecule:
+    :param atom1:
+    :param atom2:
+    :param atom3:
+    :param angle:
+    :return:
+    """
+    crd = molecule.get_atom_coordinates()
+    _, atom3_friends = molecule.divide_into_two_parts(atom2, atom3)
+    r12 = crd[molecule.atom_index[atom1]] - crd[molecule.atom_index[atom2]]
+    r23 = crd[molecule.atom_index[atom3]] - crd[molecule.atom_index[atom2]]
+    angle0 = np.arccos(np.dot(r12, r23) / np.linalg.norm(r23) / np.linalg.norm(r12))
+    delta_angle = angle - angle0
+    crd[atom3_friends] = np.dot(crd[atom3_friends] - crd[molecule.atom_index[atom2]],
+                                get_rotate_matrix(np.cross(r12, r23), delta_angle)) + crd[molecule.atom_index[atom2]]
+    for atom in molecule.atoms:
+        i = molecule.atom_index[atom]
+        atom.x = crd[i][0]
+        atom.y = crd[i][1]
+        atom.z = crd[i][2]
+
+
+def impose_dihedral(molecule, atom1, atom2, atom3, atom4, dihedral):
+    """
+
+    :param molecule:
+    :param atom1:
+    :param atom2:
+    :param atom3:
+    :param atom4:
+    :param dihedral:
+    :return:
+    """
+    crd = molecule.get_atom_coordinates()
+    _, atom4_friends = molecule.divide_into_two_parts(atom2, atom3)
+    r12 = crd[molecule.atom_index[atom1]] - crd[molecule.atom_index[atom2]]
+    r23 = crd[molecule.atom_index[atom3]] - crd[molecule.atom_index[atom2]]
+    r34 = crd[molecule.atom_index[atom3]] - crd[molecule.atom_index[atom4]]
+    r12xr23 = np.cross(r12, r23)
+    r23xr34 = np.cross(r23, r34)
+    cos = np.dot(r12xr23, r23xr34) / np.linalg.norm(r12xr23) / np.linalg.norm(r23xr34)
+    cos = max(-0.999999, min(cos, 0.999999))
+    dihedral0 = np.arccos(cos)
+    dihedral0 = np.pi - np.copysign(dihedral0, np.cross(r23xr34, r12xr23).dot(r23))
+    delta_angle = dihedral - dihedral0
[…]
