### 2022-07-19 7965985
clean ACMAP.py
