+++ Xponge/forcefield/BASE/ACMAP.py
@@ -1,10 +1,19 @@
-from ... import *
+"""
+This **module** is the basic setting of atom-specific cmap
+"""
+from ... import Generate_New_Bonded_Force_Type
+from ...helper import Molecule
 
-CMAP = Generate_New_Bonded_Force_Type("atom_specific_cmap", "1-2-3-4-5", {"resolution": int, "parameters": list}, False)
+CMap = Generate_New_Bonded_Force_Type("atom_specific_cmap", "1-2-3-4-5", {"resolution": int, "parameters": list}, False)
 
 
 @Molecule.Set_Save_SPONGE_Input("cmap")
 def write_cmap(self):
+    """
+This **function** is used to write SPONGE input file
+    :param self:
+    :return:
+    """
     bonds = []
     haved_cmaps = []
     haved_cmap_index = {}
@@ -23,7 +32,7 @@ def write_cmap(self):
                                             self.atom_index[bond.atoms[temp_order[3]]],
                                             self.atom_index[bond.atoms[temp_order[4]]], haved_cmap_index[bond.type]))
 
-    if (bonds):
+    if bonds:
         towrite = "%d %d\n" % (len(bonds), len(haved_cmaps))
         for bondtype in haved_cmaps:
             towrite += "%d " % bondtype.resolution
@@ -38,3 +47,5 @@ def write_cmap(self):
         towrite += "\n".join(bonds)
 
         return towrite
+    else:
+        return None
