+++ Xponge/forcefield/BASE/LJ.py
@@ -1,4 +1,9 @@
-from ... import *
+"""
+This **module** is the basic setting for the force field property of charge
+"""
+import numpy as np
+from ... import Generate_New_Pairwise_Force_Type
+from ...helper import Molecule, AtomType, GlobalSetting
 
 AtomType.Add_Property({"LJtype": str})
 
@@ -13,7 +18,12 @@ LJType.Set_Property_Unit("B", "energy·distance^12", "kcal/mol·A^12")
 
 
 @GlobalSetting.Add_Unit_Transfer_Function(LJType)
-def LJ_Unit_Transfer(self):
+def lj_unit_transfer(self):
+    """
+This **function** is used to transfer the units of lj
+    :param self:
+    :return:
+    """
     if self.A != None and self.B != None:
         if self.B == 0 or self.A == 0:
             self.sigma = 0
@@ -29,27 +39,49 @@ def LJ_Unit_Transfer(self):
 
 
 def Lorentz_Berthelot_For_A(epsilon1, rmin1, epsilon2, rmin2):
+    """
+This **function** is used to calculate the A coefficient for Lorentz_Berthelot mix rule
+    :param epsilon1:
+    :param rmin1:
+    :param epsilon2:
+    :param rmin2:
+    :return:
+    """
     return np.sqrt(epsilon1 * epsilon2) * ((rmin1 + rmin2) ** 12)
 
 
 def Lorents_Berthelot_For_B(epsilon1, rmin1, epsilon2, rmin2):
+    """
+This **function** is used to calculate the A coefficient for Lorentz_Berthelot mix rule
+    :param epsilon1:
+    :param rmin1:
+    :param epsilon2:
+    :param rmin2:
+    :return:
+    """
     return np.sqrt(epsilon1 * epsilon2) * 2 * ((rmin1 + rmin2) ** 6)
 
 
-def find_AB_LJ(LJtypes, stat=True):
+def _find_ab_lj(ljtypes, stat=True):
+    """
+
+    :param ljtypes:
+    :param stat:
+    :return:
+    """
     As = []
     Bs = []
 
-    for i in range(len(LJtypes)):
-        LJ_i = LJType.types[LJtypes[i] + "-" + LJtypes[i]]
+    for i in range(len(ljtypes)):
+        LJ_i = LJType.types[ljtypes[i] + "-" + ljtypes[i]]
         if stat:
-            j_max = len(LJtypes)
+            j_max = len(ljtypes)
         else:
             j_max = i + 1
         for j in range(j_max):
-            LJ_j = LJType.types[LJtypes[j] + "-" + LJtypes[j]]
+            LJ_j = LJType.types[ljtypes[j] + "-" + ljtypes[j]]
             finded = False
-            findnames = [LJtypes[i] + "-" + LJtypes[j], LJtypes[j] + "-" + LJtypes[i]]
+            findnames = [ljtypes[i] + "-" + ljtypes[j], ljtypes[j] + "-" + ljtypes[i]]
             for findname in findnames:
                 if findname in LJType.types.keys():
                     finded = True
@@ -63,13 +95,20 @@ def find_AB_LJ(LJtypes, stat=True):
     return As, Bs
 
 
-def get_checks(LJtypes, As, Bs):
+def _get_checks(ljtypes, As, Bs):
+    """
+
+    :param ljtypes:
+    :param As:
+    :param Bs:
+    :return:
+    """
     checks = {}
     count = 0
-    for i in range(len(LJtypes)):
+    for i in range(len(ljtypes)):
         check_string_A = ""
         check_string_B = ""
-        for j in range(len(LJtypes)):
+        for j in range(len(ljtypes)):
             check_string_A += "%16.7e" % As[count] + " "
[…]
