+++ Xponge/forcefield/BASE/ANGLE.py
@@ -1,4 +1,8 @@
-from ... import *
+"""
+This **module** is the basic setting for the force field format of harmonic angle
+"""
+from ... import Generate_New_Bonded_Force_Type
+from ...helper import Molecule
 
 AngleType = Generate_New_Bonded_Force_Type("angle", "1-2-3", {"k": float, "b": float}, True)
 
@@ -8,6 +12,11 @@ AngleType.Set_Property_Unit("b", "angle", "rad")
 
 @Molecule.Set_Save_SPONGE_Input("angle")
 def write_angle(self):
+    """
+This **function** is used to write SPONGE input file
+    :param self:
+    :return:
+    """
     angles = []
     for angle in self.bonded_forces.get("angle", []):
         order = list(range(3))
@@ -20,8 +29,10 @@ def write_angle(self):
                                               , self.atom_index[angle.atoms[temp_order[1]]],
                                               self.atom_index[angle.atoms[temp_order[2]]], angle.k, angle.b))
 
-    if (angles):
+    if angles:
         towrite = "%d\n" % len(angles)
         angles.sort(key=lambda x: list(map(int, x.split()[:3])))
         towrite += "\n".join(angles)
         return towrite
+
+    return None
