+++ Xponge/forcefield/BASE/BOND.py
@@ -1,4 +1,8 @@
-from ... import *
+"""
+This **module** is the basic setting for the force field format of harmonic bond
+"""
+from ... import Generate_New_Bonded_Force_Type
+from ...helper import Molecule
 
 BondType = Generate_New_Bonded_Force_Type("bond", "1-2", {"k": float, "b": float}, True)
 
@@ -8,6 +12,11 @@ BondType.Set_Property_Unit("b", "distance", "A")
 
 @Molecule.Set_Save_SPONGE_Input("bond")
 def write_bond(self):
+    """
+This **function** is used to write SPONGE input file
+    :param self:
+    :return:
+    """
     bonds = []
     for bond in self.bonded_forces.get("bond", []):
         order = list(range(2))
@@ -19,9 +28,10 @@ def write_bond(self):
             bonds.append("%d %d %f %f" % (self.atom_index[bond.atoms[temp_order[0]]]
                                           , self.atom_index[bond.atoms[temp_order[1]]], bond.k, bond.b))
 
-    if (bonds):
+    if bonds:
         towrite = "%d\n" % len(bonds)
         bonds.sort(key=lambda x: list(map(int, x.split()[:2])))
         towrite += "\n".join(bonds)
 
         return towrite
+    return None
