+++ Xponge/forcefield/BASE/VATOM.py
@@ -1,4 +1,8 @@
-from ... import *
+"""
+This **module** is the basic setting for the force field format of virtual atoms
+"""
+from ... import Generate_New_Bonded_Force_Type
+from ...helper import Molecule, GlobalSetting
 
 VirtualType2 = Generate_New_Bonded_Force_Type("vatom2", "1",
                                               {"atom0": int, "atom1": int, "atom2": int, "k1": float, "k2": float},
@@ -8,6 +12,11 @@ GlobalSetting.VirtualAtomTypes["vatom2"] = 3
 
 @Molecule.Set_Save_SPONGE_Input("virtual_atom")
 def write_virtual_atoms(self):
+    """
+This **function** is used to write SPONGE input file
+    :param self:
+    :return:
+    """
     vatoms = []
     for vatom in self.bonded_forces.get("vatom2", []):
         vatoms.append("2 %d %d %d %d %f %f" % (self.atom_index[vatom.atoms[0]],
@@ -16,8 +25,9 @@ def write_virtual_atoms(self):
                                                self.atom_index[vatom.atoms[0]] + vatom.atom2,
                                                vatom.k1, vatom.k2))
 
-    if (vatoms):
+    if vatoms:
         towrite = ""
         towrite += "\n".join(vatoms)
 
         return towrite
+    return None
