+++ Xponge/forcefield/BASE/RCMAP.py
@@ -1,30 +1,39 @@
-from ... import *
+"""
+This **module** is the basic setting for the force field format of atom-specific cmap
+"""
+from ... import Generate_New_Bonded_Force_Type
+from ...helper import Molecule
 
-CMAP = Generate_New_Bonded_Force_Type("residue_specific_cmap", "1-2-3-4-5", {}, False)
+CMapType = Generate_New_Bonded_Force_Type("residue_specific_cmap", "1-2-3-4-5", {}, False)
 
-CMAP.Residue_Map = {}
+CMapType.Residue_Map = {}
 
 
 @Molecule.Set_Save_SPONGE_Input("cmap")
 def write_cmap(self):
+    """
+This **function** is used to write SPONGE input file
+    :param self:
+    :return:
+    """
     cmaps = []
     resolutions = []
     used_types = []
     used_types_map = {}
-    atoms = []
+    cmap = None
     for cmap in self.bonded_forces.get("residue_specific_cmap", []):
         resname = cmap.atoms[2].residue.type.name
-        if resname in CMAP.Residue_Map.keys():
-            if CMAP.Residue_Map[resname]["count"] not in used_types_map.keys():
-                used_types_map[CMAP.Residue_Map[resname]["count"]] = len(used_types)
-                used_types.append(CMAP.Residue_Map[resname]["parameters"])
-                resolutions.append(str(CMAP.Residue_Map[resname]["resolution"]))
+        if resname in CMapType.Residue_Map.keys():
+            if CMapType.Residue_Map[resname]["count"] not in used_types_map.keys():
+                used_types_map[CMapType.Residue_Map[resname]["count"]] = len(used_types)
+                used_types.append(CMapType.Residue_Map[resname]["parameters"])
+                resolutions.append(str(CMapType.Residue_Map[resname]["resolution"]))
             cmaps.append("%d %d %d %d %d %d" % (self.atom_index[cmap.atoms[0]], self.atom_index[cmap.atoms[1]],
                                                 self.atom_index[cmap.atoms[2]], self.atom_index[cmap.atoms[3]],
                                                 self.atom_index[cmap.atoms[4]],
-                                                used_types_map[CMAP.Residue_Map[resname]["count"]]))
+                                                used_types_map[CMapType.Residue_Map[resname]["count"]]))
 
-    if (cmap):
+    if cmap:
         towrite = "%d %d\n" % (len(cmaps), len(resolutions))
         towrite += " ".join(resolutions) + "\n\n"
 
@@ -39,3 +48,4 @@ def write_cmap(self):
         towrite += "\n".join(cmaps)
 
         return towrite
+    return None
