+++ Xponge/forcefield/AMBER/gaff.py
@@ -12,25 +12,25 @@ DIHEDRAL.ProperType.New_From_String(propers)
 DIHEDRAL.ImproperType.New_From_String(impropers)
 LJ.LJType.New_From_String(LJs)
 
-GAFF = assign.Judge_Rule("GAFF")
+GAFF = assign.AssignRule("GAFF")
 
 
-@GAFF.Add_Judge_Rule("cx")
+@GAFF.Add_Rule("cx")
 def temp(i, Assign):
     return Assign.Atom_Judge(i, "C4") and "RG3" in Assign.atom_marker[i].keys()
 
 
-@GAFF.Add_Judge_Rule("cy")
+@GAFF.Add_Rule("cy")
 def temp(i, Assign):
     return Assign.Atom_Judge(i, "C4") and "RG4" in Assign.atom_marker[i].keys()
 
 
-@GAFF.Add_Judge_Rule("c3")
+@GAFF.Add_Rule("c3")
 def temp(i, Assign):
     return Assign.Atom_Judge(i, "C4")
 
 
-@GAFF.Add_Judge_Rule("c")
+@GAFF.Add_Rule("c")
 def temp(i, Assign):
     tofind = Assign.Atom_Judge(i, "C3")
     if tofind:
@@ -42,7 +42,7 @@ def temp(i, Assign):
     return tofind
 
 
-@GAFF.Add_Judge_Rule("cz")
+@GAFF.Add_Rule("cz")
 def temp(i, Assign):
     tofind = Assign.Atom_Judge(i, "C3")
     if tofind:
@@ -53,7 +53,7 @@ def temp(i, Assign):
     return tofind
 
 
-@GAFF.Add_Judge_Rule("cq")
+@GAFF.Add_Rule("cq")
 def temp(i, Assign):
     tofind = Assign.Atom_Judge(i, "C3") and "AR1" in Assign.atom_marker[i] and Assign.atom_marker[i].get("RG6") == 1
     for bonded_atom in Assign.bonds[i].keys():
@@ -72,7 +72,7 @@ def temp(i, Assign):
     return tofind
 
 
-@GAFF.Add_Judge_Rule("cp")
+@GAFF.Add_Rule("cp")
 def temp(i, Assign):
     tofind = Assign.Atom_Judge(i, "C3") and "AR1" in Assign.atom_marker[i] and Assign.atom_marker[i].get("RG6") == 1
     for bonded_atom in Assign.bonds[i].keys():
@@ -84,7 +84,7 @@ def temp(i, Assign):
     return tofind
 
 
-@GAFF.Add_Judge_Rule("ca")
+@GAFF.Add_Rule("ca")
 def temp(i, Assign):
     return Assign.Atom_Judge(i, "C3") and "AR1" in Assign.atom_marker[i]
 
@@ -101,7 +101,7 @@ def _single_double_name(i, Assign, name1, name2):
     return tofind
 
 
-@GAFF.Add_Judge_Rule("cd")
+@GAFF.Add_Rule("cd")
 def temp(i, Assign):
     tofind = Assign.Atom_Judge(i, "C3") and "sb" in Assign.atom_marker[i] and "db" in Assign.atom_marker[i] and (
                 "AR2" in Assign.atom_marker[i] or "AR3" in Assign.atom_marker[i])
@@ -110,13 +110,13 @@ def temp(i, Assign):
     return tofind
 
 
-@GAFF.Add_Judge_Rule("cc")
+@GAFF.Add_Rule("cc")
 def temp(i, Assign):
     return Assign.Atom_Judge(i, "C3") and "sb" in Assign.atom_marker[i] and "db" in Assign.atom_marker[i] and (
[…]
