+++ Xponge/assign/__init__.py
@@ -1,20 +1,38 @@
-from .. import *
-import pubchempy as pcp
+"""
+This **package** is used to assign the properties for atoms, residues and molecules
+"""
+# from .. import *
+# import pubchempy as pcp
+# from itertools import groupby
+# import sys
+from collections import OrderedDict
 from itertools import groupby
-import sys
+import numpy as np
+from ..helper import set_attribute_alternative_names, AtomType, guess_element_from_mass
 
 
-class Judge_Rule:
+class AssignRule:
+    """
+This **class** is to be the rule to determine the atom type for one atom
+    """
     all = {}
 
     def __init__(self, name):
         self.name = name
-        Judge_Rule.all[name] = self
+        AssignRule.all[name] = self
         self.rules = OrderedDict()
+        set_attribute_alternative_names(self)
 
-    def Add_Judge_Rule(self, atomtype):
+    def add_rule(self, atomtype):
+        """
+This **function** is used as an **decorator** to add the atom type - judge function
+        :param atomtype:
+        :return:
+        """
         if isinstance(atomtype, str):
             atomtype = AtomType.types[atomtype]
+        elif not isinstance(atomtype, AtomType):
+            raise TypeError("atomtype should be a string or AtomType")
 
         def wrapper(rule_function):
             self.rules[atomtype] = rule_function
@@ -23,15 +41,31 @@ class Judge_Rule:
 
 
 class _RING():
+    """
+This **class** is used to help with the ring assignment.
+    """
+
     @staticmethod
-    def Add_Rings_Basic_Marker(assign, rings):
+    def add_rings_basic_marker(assign, rings):
+        """
+
+        :param assign:
+        :param rings:
+        :return:
+        """
         for ring in rings:
             for atom in ring.atoms:
                 assign.Add_Atom_Marker(atom, "RG")
                 assign.Add_Atom_Marker(atom, "RG%d" % len(ring.atoms))
 
     @staticmethod
-    def Check_Rings_Type(assign, rings):
+    def check_rings_type(assign, rings):
+        """
+
+        :param assign:
+        :param rings:
+        :return:
+        """
         for ring in rings:
             ring.check_pure_aromatic(assign)
             ring.check_pure_aliphatic_and_planar(assign)
@@ -50,7 +84,12 @@ class _RING():
                         assign.Add_Atom_Marker(atom, "AR4")
 
     @staticmethod
-    def Get_Rings(assign):
+    def get_rings(assign):
+        """
+
+        :param assign:
+        :return:
+        """
         current_path = []
         current_path_sons = {}
         current_work = []
@@ -97,7 +136,7 @@ class _RING():
         return hash(self.tohash)
 
     def __eq__(self, other):
-        return type(other) == _RING and self.tohash == other.tohash
+        return isinstance(other, _RING) and self.tohash == other.tohash
 
     def __init__(self, atom_list):
         min_index = np.argmin(atom_list)
@@ -109,10 +148,19 @@ class _RING():
         self.tohash = "-".join(["%d" % atom for atom in self.atoms])
 
     def get_3_neighbors(self):
-        for i in range(len(self.atoms)):
-            yield self.atoms[i - 2], self.atoms[i - 1], self.atoms[i]
+        """
+
+        :return:
+        """
+        for i, atom in enumerate(self.atoms):
+            yield self.atoms[i - 2], self.atoms[i - 1], atom
 
     def check_pure_aromatic(self, assign):
+        """
+
+        :param assign:
+        :return:
+        """
[…]

[…]
