+++ Xponge/tools/__init__.py
@@ -35,15 +35,28 @@ def assign_test(args):
 
     globals().update(__main__.__dict__)
     t = assign.Assign()
-    t.add_atom("O", 0,0,0)
-    t.add_atom("H", 1,0,0)
-    t.add_atom("H", 0,1,0)
-    t.add_bond(0,1,1)
-    t.add_bond(0,2,1)
+    t.add_atom("O", 0, 0, 0)
+    t.add_atom("H", 1, 0, 0)
+    t.add_atom("H", 0, 1, 0)
+    t.add_bond(0, 1, 1)
+    t.add_bond(0, 2, 1)
     t.determine_ring_and_bond_type()
     t.determine_atom_type("GAFF")
+    t.calculate_charge("resp", opt = True)
+    Save_PDB(t, f"{args.o}.pdb")
+    Save_Mol2(t, f"{args.o}.mol2")
+    WAT = t.to_residuetype("WAT")
+    Save_PDB(WAT, f"{args.o}_Residue.pdb")
+    Save_Mol2(WAT, f"{args.o}_Residue.pdb")
+    Save_SPONGE_Input(WAT, f"{args.o}")
 
+def charmm27_test(args):
+    import Xponge
+    import Xponge.forcefield.CHARMM27.protein
+    globals().update(__main__.__dict__)
 
+    t = ACE + ALA * 10 + NME
+    Save_SPONGE_Input(t, f"{args.o}")
 
 def test(args):
     if not args.do:
@@ -54,6 +67,9 @@ def test(args):
 
     if "assign" in args.do:
         assign_test(args)
+
+    if "charmm27" in args.do:
+        charmm27_test(args)
 
 
 def dat2nc(args):
