+++ Xponge/assign/__init__.py
@@ -1,14 +1,11 @@
 """
 This **package** is used to assign the properties for atoms, residues and molecules
 """
-# from .. import *
-# import pubchempy as pcp
-# from itertools import groupby
-# import sys
+import sys
 from collections import OrderedDict
 from itertools import groupby
 import numpy as np
-from ..helper import set_attribute_alternative_names, AtomType, guess_element_from_mass
+from ..helper import set_attribute_alternative_names, AtomType, guess_element_from_mass, ResidueType
 
 
 class AssignRule:
@@ -464,6 +461,7 @@ This **class** is used to assign properties for atoms, which is called an "assig
         :param parameters:
         :return:
         """
+        method = method.upper()
         if method == "RESP":
             from . import RESP
             self.charge = RESP.RESP_Fit(self, basis=parameters.get("basis", "6-31g*"), opt=parameters.get("opt", False),
@@ -473,8 +471,10 @@ This **class** is used to assign properties for atoms, which is called an "assig
                                         grid_cell_layer=parameters.get("grid_cell_layer", 4),
                                         a1=parameters.get("a1", 0.0005),
                                         a2=parameters.get("a2", 0.001), two_stage=parameters.get("two_stage", True),
-                                        only_ESP=parameters.get("only_ESP", False),
+                                        only_esp=parameters.get("only_esp", False),
                                         radius=parameters.get("radius", None))
+        else:
+            raise ValueError("methods should be one of the following: 'RESP'")
 
     def save_as_pdb(self, filename):
         """
@@ -482,6 +482,8 @@ This **class** is used to assign properties for atoms, which is called an "assig
         :param filename:
         :return:
         """
+        if not isinstance(filename, str):
+            raise TypeError("filename needed to save an assignment to a pdb file")
         towrite = towrite = "REMARK   Generated By Xponge (Assignment)\n"
         count = {}
         for i in range(self.atom_numbers):
@@ -522,6 +524,8 @@ This **class** is used to assign properties for atoms, which is called an "assig
         :param filename:
         :return:
         """
+        if not isinstance(filename, str):
+            raise TypeError("filename needed to save an assignment as a mol2 file")
         bonds = []
         for i in range(self.atom_numbers):
             for j, order in self.bonds[i].items():
