+++ setup.py
@@ -7,7 +7,7 @@ f.close()
 #for formal
 setuptools.setup(
     name="Xponge",
-    version="1.2.5.17",
+    version="1.2.6.0",
     author="Yijie Xia",  
     author_email="yijiexia@pku.edu.cn", 
     description="A package for building molecular dynamics inputs for SPONGE",
@@ -16,7 +16,7 @@ setuptools.setup(
     url="https://gitee.com/gao_hyp_xyj_admin/xponge",
     packages=setuptools.find_packages(),
     package_data = {"":['*.mol2', '*.frcmod', '*.dat', '*.itp']},
-    install_requires = ["numpy", "pubchempy", "NetCDF4"],
+    install_requires = ["numpy"],
     entry_points = {
         "console_scripts": ["Xponge = Xponge.__main__:main"] },
     classifiers=[
