+++ Xponge.egg-info/SOURCES.txt
@@ -2,11 +2,14 @@ LICENSE
 README.md
 setup.py
 Xponge/BUILD.py
-Xponge/IMPOSE.py
 Xponge/LOAD.py
 Xponge/PROCESS.py
 Xponge/__init__.py
 Xponge/__main__.py
+Xponge/build.py
+Xponge/helper.py
+Xponge/load.py
+Xponge/process.py
 Xponge.egg-info/PKG-INFO
 Xponge.egg-info/SOURCES.txt
 Xponge.egg-info/dependency_links.txt
@@ -65,11 +68,11 @@ Xponge/forcefield/BASE/IMPROPER.py
 Xponge/forcefield/BASE/LJ.py
 Xponge/forcefield/BASE/MASS.py
 Xponge/forcefield/BASE/NB14.py
-Xponge/forcefield/BASE/NB14_EXTRA.py
+Xponge/forcefield/BASE/NB14EXTRA.py
 Xponge/forcefield/BASE/RCMAP.py
-Xponge/forcefield/BASE/SOFT_BOND.py
-Xponge/forcefield/BASE/UREY_BRADLEY.py
-Xponge/forcefield/BASE/VIRTUAL_ATOM.py
+Xponge/forcefield/BASE/SBOND.py
+Xponge/forcefield/BASE/UBANGLE.py
+Xponge/forcefield/BASE/VATOM.py
 Xponge/forcefield/BASE/__init__.py
 Xponge/forcefield/CHARMM27/__init__.py
 Xponge/forcefield/CHARMM27/atomic_ions.mol2
@@ -83,6 +86,7 @@ Xponge/forcefield/CHARMM27/tip3p.mol2
 Xponge/forcefield/CHARMM27/tip3p.py
 Xponge/forcefield/CHARMM27/tips3p.py
 Xponge/forcefield/SPECIAL/FEP.py
+Xponge/forcefield/SPECIAL/GB.py
 Xponge/forcefield/SPECIAL/MIN_BONDED.py
 Xponge/forcefield/SPECIAL/__init__.py
 Xponge/tools/__init__.py
