+++ Xponge.egg-info/requires.txt
@@ -1,3 +1,1 @@
 numpy
-pubchempy
-NetCDF4

[…]
