### 2022-07-21 2ebf5f9
clean __main__.py
