+++ Xponge/__main__.py
@@ -1,22 +1,37 @@
+"""
+This **module** defines the terminal commands
+"""
+import argparse
 from . import tools
 
 
 def _mytest(subparsers):
+    """
+
+    :param subparsers:
+    :return:
+    """
     mytest = subparsers.add_parser("test", help="test the basic function of Xponge")
     mytest.add_argument("-o", metavar="test", default="test", help="the prefix for the output files")
     mytest.add_argument("-do", metavar="todo", nargs="*", action="append",
-                        default = None, help="the things need to test, should be one or more of 'base', 'assign', 'charmm27'",
-                        choices=["base", "assign", "charmm27"])
+                        default=None, choices=["base", "assign", "charmm27"],
+                        help="the things need to test, should be one or more of 'base', 'assign', 'charmm27'")
     mytest.set_defaults(func=tools.test)
 
 
 def _dat2nc(subparsers):
+    """
+
+    :param subparsers:
+    :return:
+    """
     dat2nc = subparsers.add_parser("dat2nc", help="convert a traj file from .dat to .nc")
 
     dat2nc.add_argument("-n", required=True, type=int, help="the number of atoms in the traj file",
                         metavar="atom_numbers")
     dat2nc.add_argument("-frame", type=int,
-                        help="the number of frames you want convert. If not set, a value will be guessed from the file size.",
+                        help="the number of frames you want convert. If not set, \
+                            a value will be guessed from the file size.",
                         metavar="frame_numbers")
     dat2nc.add_argument("-x", required=True, help="the name of the .dat traj file.", metavar="dat_traj_file")
     group = dat2nc.add_mutually_exclusive_group(required=True)
@@ -30,6 +45,11 @@ def _dat2nc(subparsers):
 
 
 def _gro2crd(subparsers):
+    """
+
+    :param subparsers:
+    :return:
+    """
     gro2crd = subparsers.add_parser("gro2crd", help="convert a coordinate file from gromacs to SPONGE")
     gro2crd.add_argument("-i", required=True, help="the name of the .gro file.")
     gro2crd.add_argument("-o", required=True, help="the name of the SPONGE coordinate file.")
@@ -37,6 +57,11 @@ def _gro2crd(subparsers):
 
 
 def _nc2rst7(subparsers):
+    """
+
+    :param subparsers:
+    :return:
+    """
     nc2rst7 = subparsers.add_parser("nc2rst7", help="convert a rst7 file from .nc to .rst7")
     nc2rst7.add_argument("-rst7", required=True, help="the name of the .rst7 restart file.")
     nc2rst7.add_argument("-nc", required=True, help="the name of the .nc restart file.")
@@ -44,8 +69,13 @@ def _nc2rst7(subparsers):
 
 
 def _maskgen(subparsers):
+    """
+
+    :param subparsers:
+    :return:
+    """
     maskgen = subparsers.add_parser("maskgen",
-                                    help="use VMD to generate a file to record the atom indexes of the corresponding mask")
+        help="use VMD to generate a file to record the atom indexes of the corresponding mask")
     maskgen.add_argument("-p", required=True, help="the topology file")
     maskgen.add_argument("-c", help="the coordinate file")
     maskgen.add_argument("-o", required=True, help="the output file")
@@ -54,24 +84,34 @@ def _maskgen(subparsers):
 
 
 def _exgen(subparsers):
+    """
+
+    :param subparsers:
+    :return:
+    """
     exgen = subparsers.add_parser("exgen",
[…]
