### 2022-07-21 6b18310
clean rdkit tools
