+++ Xponge/forcefield/base/lj_base.py
@@ -1,0 +1,206 @@
+"""
+This **module** is the basic setting for the force field property of charge
+"""
+import numpy as np
+from ... import Generate_New_Pairwise_Force_Type
+from ...helper import Molecule, AtomType, GlobalSetting, set_dict_value_alternative_name
+
+AtomType.Add_Property({"LJtype": str})
+
+# pylint: disable=invalid-name
+LJType = Generate_New_Pairwise_Force_Type("LJ",
+                                          {"epsilon": float, "rmin": float, "sigma": float, "A": float, "B": float})
+
+LJType.Set_Property_Unit("rmin", "distance", "A")
+LJType.Set_Property_Unit("sigma", "distance", "A")
+LJType.Set_Property_Unit("epsilon", "energy", "kcal/mol")
+LJType.Set_Property_Unit("A", "energy·distance^6", "kcal/mol·A^6")
+LJType.Set_Property_Unit("B", "energy·distance^12", "kcal/mol·A^12")
+
+
+@GlobalSetting.Add_Unit_Transfer_Function(LJType)
+def lj_unit_transfer(self):
+    """
+This **function** is used to transfer the units of lj
+    :param self:
+    :return:
+    """
+    if self.A is not None and self.B is not None:
+        if self.B == 0 or self.A == 0:
+            self.sigma = 0
+            self.epsilon = 0
+        else:
+            self.sigma = (self.A / self.B) ** (1 / 6)
+            self.epsilon = 0.25 * self.B * self.sigma ** (-6)
+        self.A = None
+        self.B = None
+    if self.sigma is not None:
+        self.rmin = self.sigma * (4 ** (1 / 12) / 2)
+        self.sigma = None
+
+
+def lorentz_berthelot_for_a(epsilon1, rmin1, epsilon2, rmin2):
+    """
+This **function** is used to calculate the A coefficient for Lorentz_Berthelot mix rule
+    :param epsilon1:
+    :param rmin1:
+    :param epsilon2:
+    :param rmin2:
+    :return:
+    """
+    return np.sqrt(epsilon1 * epsilon2) * ((rmin1 + rmin2) ** 12)
+
+
+set_dict_value_alternative_name(globals(), lorentz_berthelot_for_a)
+
+
+def lorentz_berthelot_for_b(epsilon1, rmin1, epsilon2, rmin2):
+    """
+This **function** is used to calculate the A coefficient for Lorentz_Berthelot mix rule
+    :param epsilon1:
+    :param rmin1:
+    :param epsilon2:
+    :param rmin2:
+    :return:
+    """
+    return np.sqrt(epsilon1 * epsilon2) * 2 * ((rmin1 + rmin2) ** 6)
+
+
+set_dict_value_alternative_name(globals(), lorentz_berthelot_for_b)
+
+
+def _find_ab_lj(ljtypes, stat=True):
+    """
+
+    :param ljtypes:
+    :param stat:
+    :return:
+    """
+    coefficients_a = []
+    coefficients_b = []
+
+    for i in range(len(ljtypes)):
+        lj_i = LJType.types[ljtypes[i] + "-" + ljtypes[i]]
+        if stat:
+            j_max = len(ljtypes)
+        else:
+            j_max = i + 1
+        for j in range(j_max):
+            lj_j = LJType.types[ljtypes[j] + "-" + ljtypes[j]]
+            finded = False
+            findnames = [ljtypes[i] + "-" + ljtypes[j], ljtypes[j] + "-" + ljtypes[i]]
+            for findname in findnames:
+                if findname in LJType.types.keys():
+                    finded = True
+                    lj_ij = LJType.types[findname]
+                    coefficients_a.append(
[…]
