+++ Xponge/forcefield/special/fep.py
@@ -1,11 +1,14 @@
-from ... import *
-from ..base import lj_base
-from ..base import exclude_base
-from ..base import bond_base
+"""
+This **module** gives the basic functions for fep calculations
+"""
+from ...helper import source, set_global_alternative_names
+from ..base import lj_base, exclude_base, bond_base, angle_base, dihedral_base, nb14_extra_base
 
-LJType = LJ.LJType
+source("....")
 
-LJType.New_From_String("""name A  B
+lj_type = lj_base.LJType
+
+lj_type.New_From_String("""name A  B
 ZERO_LJ_ATOM-ZERO_LJ_ATOM  0  0
 """)
 
@@ -13,223 +16,312 @@ AtomType.Add_Property({"LJtypeB": str})
 AtomType.Add_Property({"subsys": int})
 
 
-def _find_common_forces(forcetype, Aforces, Bforces, molB2molA):
+def _find_common_forces(forcetype, aforces, bforces, mol_b2mol_a):
+    """
+
+    :param forcetype:
+    :param aforces:
+    :param bforces:
+    :param mol_b2mol_a:
+    :return:
+    """
     toret = []
     temp_map = {}
     temp_map2 = {}
-    for force in Bforces:
+    for force in bforces:
         temp_map2[force] = True
         for fatoms in forcetype.Same_Force(force.atoms):
-            temp_map["-".join(list(map(lambda atom: str(molB2molA[atom]), fatoms)))] = force
-    for force in Aforces:
+            temp_map["-".join(list(map(lambda atom: str(mol_b2mol_a[atom]), fatoms)))] = force
+    for force in aforces:
         tofind = "-".join(list(map(str, force.atoms)))
-        Bforce = temp_map.get(tofind, None)
-        toret.append([force, Bforce])
-        temp_map2[Bforce] = False
-    for force in Bforces:
+        bforce = temp_map.get(tofind, None)
+        toret.append([force, bforce])
+        temp_map2[bforce] = False
+    for force in bforces:
         if temp_map2[force]:
             toret.append([None, force])
     return toret
 
 
-FEP_Bonded_Force_Merge_Rule = {}
+FEP_BONDED_FORCE_MERGE_RULE = {}
+TINY = 1e-10 / 18.2223
 
 
-def nb14_extra_merge_rule(molR, molA, molB, forcetype, Rforces, Bforces, _lambda, molR2molA, molA2molR, molR2molB,
-                          molB2molR):
-    TINY = 1e-10 / 18.2223
-    forcepair = _find_common_forces(forcetype, Rforces, Bforces, molB2molR)
-    for fR, fB in forcepair:
-        if fB is None:
-            temp_charge0 = fR.atoms[0].charge if abs(fR.atoms[0].charge) > TINY else TINY
-            temp_charge1 = fR.atoms[1].charge if abs(fR.atoms[1].charge) > TINY else TINY
+# pylint: disable=unused-argument
+def _nb14_extra_merge_rule(mol_r, mol_a, mol_b, forcetype, rforces, bforces, lambda_, mol_r2mol_a, mol_a2mol_r,
+                           mol_r2mol_b, mol_b2mol_r):
+    """
 
-            if fR.nb14_ee_factor is not None:
-                nb14_ee_factor = fR.nb14_ee_factor
+    :param mol_r:
+    :param mol_a:
+    :param mol_b:
+    :param forcetype:
+    :param rforces:
+    :param bforces:
+    :param lambda_:
+    :param mol_r2mol_a:
+    :param mol_a2mol_r:
+    :param mol_r2mol_b:
+    :param mol_b2mol_r:
+    :return:
[…]
