+++ Xponge/forcefield/amber/spce.py
@@ -25,7 +25,7 @@ HW-OW-HW  0   0
 OW-HW-HW  0   0
 """)
 
-lj_base.LJType.New_From_String(r"""
+lj_base.lj_type.New_From_String(r"""
 name    epsilon[kcal/mol]   rmin[A]
 OW-OW   0.1553              1.7767
 HW-HW   0                   0
