+++ Xponge/forcefield/charmm27/tip3p.py
@@ -26,7 +26,7 @@ HW-OW-HW  0   0
 OW-HW-HW  0   0
 """)
 
-lj_base.LJType.New_From_String(r"""
+lj_base.lj_type.New_From_String(r"""
 name    epsilon[kcal/mol]   rmin[A]
 OW-OW   0.152               1.7683
 HW-HW   0                   0
