+++ Xponge/forcefield/charmm27/tips3p.py
@@ -25,7 +25,7 @@ HW-OW-HW  0   0
 OW-HW-HW  0   0
 """)
 
-lj_base.LJType.New_From_String(r"""
+lj_base.lj_type.New_From_String(r"""
 name    sigma[nm]   epsilon[kJ/mol]
 OW-OW   0.315057422683    0.6363864
 HW-HW   0.0400013524445 0.192464
