+++ Xponge/forcefield/charmm27/protein.py
@@ -8,9 +8,9 @@ source("...charmm27")
 
 load_mol2(os.path.join(CHARMM27_DATA_DIR, "protein.mol2"))
 
-ResidueType.types.pop("HYP")
-ResidueType.types.pop("CHYP")
-ResidueType.types.pop("NHE")
+ResidueType.clear_type("HYP")
+ResidueType.clear_type("CHYP")
+ResidueType.clear_type("NHE")
 
 remove_real_global_variable("HYP")
 remove_real_global_variable("CHYP")
@@ -19,37 +19,37 @@ remove_real_global_variable("NHE")
 residues = "ALA ARG ASN ASP CYS CYX GLN GLU GLY HID HIE HIP ILE LEU LYS MET PHE PRO SER THR TRP TYR VAL".split()
 
 for res in residues:
-    ResidueType.types[res].head_next = "CA"
-    ResidueType.types[res].head_length = 1.3
-    ResidueType.types[res].tail_next = "CA"
-    ResidueType.types[res].tail_length = 1.3
-    ResidueType.types[res].head_link_conditions.append({"atoms": ["CA", "N"], "parameter": 120 / 180 * np.pi})
-    ResidueType.types[res].head_link_conditions.append({"atoms": ["H", "CA", "N"], "parameter": -np.pi})
-    ResidueType.types[res].tail_link_conditions.append({"atoms": ["CA", "C"], "parameter": 120 / 180 * np.pi})
-    ResidueType.types[res].tail_link_conditions.append({"atoms": ["O", "CA", "C"], "parameter": -np.pi})
+    ResidueType.get_type(res).head_next = "CA"
+    ResidueType.get_type(res).head_length = 1.3
+    ResidueType.get_type(res).tail_next = "CA"
+    ResidueType.get_type(res).tail_length = 1.3
+    ResidueType.get_type(res).head_link_conditions.append({"atoms": ["CA", "N"], "parameter": 120 / 180 * np.pi})
+    ResidueType.get_type(res).head_link_conditions.append({"atoms": ["H", "CA", "N"], "parameter": -np.pi})
+    ResidueType.get_type(res).tail_link_conditions.append({"atoms": ["CA", "C"], "parameter": 120 / 180 * np.pi})
+    ResidueType.get_type(res).tail_link_conditions.append({"atoms": ["O", "CA", "C"], "parameter": -np.pi})
 
-    ResidueType.types["N" + res].tail_next = "CA"
-    ResidueType.types["N" + res].tail_length = 1.3
-    ResidueType.types["N" + res].tail_link_conditions.append({"atoms": ["CA", "C"], "parameter": 120 / 180 * np.pi})
-    ResidueType.types["N" + res].tail_link_conditions.append({"atoms": ["O", "CA", "C"], "parameter": -np.pi})
+    ResidueType.get_type("N" + res).tail_next = "CA"
+    ResidueType.get_type("N" + res).tail_length = 1.3
+    ResidueType.get_type("N" + res).tail_link_conditions.append({"atoms": ["CA", "C"], "parameter": 120 / 180 * np.pi})
+    ResidueType.get_type("N" + res).tail_link_conditions.append({"atoms": ["O", "CA", "C"], "parameter": -np.pi})
 
-    ResidueType.types["C" + res].head_next = "CA"
-    ResidueType.types["C" + res].head_length = 1.3
-    ResidueType.types["C" + res].head_link_conditions.append({"atoms": ["CA", "N"], "parameter": 120 / 180 * np.pi})
[…]
