+++ Xponge/forcefield/amber/gaff.py
@@ -98,7 +98,7 @@ def _rule_cq(i, assign):
         for bonded_atom in assign.bonds[i]:
             if tofind:
                 break
-            if (assign.atom_types[bonded_atom] == AtomType.types["cp"] and \
+            if (assign.atom_types[bonded_atom] == AtomType.get_type("cp") and \
                     "AB" in assign.bond_marker[bonded_atom][i]):
                 tofind = True
                 break
@@ -147,8 +147,8 @@ def _single_double_name(i, assign, name1, name2):
     for bonded_atom, bond_order in assign.bonds[i].items():
         if tofind:
             break
-        if ((assign.atom_types[bonded_atom] == AtomType.types[name1] and bond_order == 2) or
-                (assign.atom_types[bonded_atom] == AtomType.types[name2] and bond_order == 1)):
+        if ((assign.atom_types[bonded_atom] == AtomType.get_type(name1) and bond_order == 2) or
+                (assign.atom_types[bonded_atom] == AtomType.get_type(name2) and bond_order == 1)):
             tofind = True
             break
     return tofind
@@ -1272,7 +1272,7 @@ def _rule_p5(i, assign):
 
 def parmchk2_gaff(ifname, ofname, direct_load=True, keep=True):
     """
-This **function** is to do parmchk2 for gaff
+    This **function** is to do parmchk2 for gaff
     :param ifname:
     :param ofname:
     :param direct_load:
@@ -1297,3 +1297,4 @@ Xprint("""Reference for gaff:
     Journal of Computational Chemistry 2004 25, 1157-1174
     DOI: 10.1002/jcc.20035
 """)
+# pylint:disable=undefined-variable
