### 2022-08-05 Xia-Yijie/Xponge#1 @ojeda-e
[JOSS review]:  Xponge package has no automated tests
As part of the JOSS review, the Xponge package should include automated tests.

Description
----------------
Currently, the Xponge repo has a [tests_standard/](https://github.com/Xia-Yijie/Xponge/tree/master/test_standard) folder that includes files with the `test_` prefix. These files are, however, text files instead of `.py` files with unit tests.
With no automated tests, the functionality of the code cannot be assessed.

Suggested approach
----------------------------
To fix this issue, please add automated test to verify functionality of your code. Please check the documentation of the [Pytest](https://docs.pytest.org/en/7.1.x/) suite. Alternatively you can check [Unittest](https://docs.python.org/3/library/unittest.html) framework.

Documentation should include instructions to run Xponge tests locally.
