### 2022-08-05 Xia-Yijie/Xponge#2 @ojeda-e
[JOSS review]: Installation from source fails
Description
------------------
In the README, Xponge offers instruction to install with pip and from source. For the latter one, the steps provided read as:

> download or clone the source of this repository
> 2.2 open the directory where you download or clone the repository
> 2.3 run the command
> `python setup.py install`
> After the installation, you can use the following command to test whether your installation is successful.
> `python -m Xponge test -o test`
> or just
> `Xponge test`

Following these steps, installation from source fails on Ubuntu 20.04. For example, I got the error:
```
Traceback (most recent call last):
  File "xponge-del/lib/python3.10/runpy.py", line 196, in _run_module_as_main
    return _run_code(code, main_globals, None,
  File "xponge-del/lib/python3.10/runpy.py", line 86, in _run_code
    exec(code, run_globals)
  File "review/Xponge/Xponge/__main__.py", line 194, in <module>
    main()
  File "review/Xponge/Xponge/__main__.py", line 189, in main
    args.func(args)
  File "review/Xponge/Xponge/tools/__init__.py", line 22, in test
    Process_Box(t, WAT, 10)
  File "review/Xponge/Xponge/PROCESS.py", line 66, in Box
    _Box(molecule, new_molecule, molmin, solshape, distance)
  File "review/Xponge/Xponge/PROCESS.py", line 6, in _Box
    while x0 < molmax[0] + distance[3] + solshape[0]:
NameError: name 'molmax' is not defined. Did you mean: 'molmin'?
```

Questions
---------------
@Xia-Yijie could you please explain the functionality of lines 20-21 of your setup.py?
https://github.com/Xia-Yijie/Xponge/blob/ae3555b0c327d6c1d8cc18952b008a49b52d9b00/setup.py#L20-L21

As far as I can tell, the installation issue is originated in that entry point.

Other suggestions
---------------------------
Additionally, if you already know the dependencies of Xponge, it would be better to provide a `requirements.yaml` file to create a new environment with the list of conda and pip. With that file installation from source can be simplified to:
```
git clone git@github.com:Xia-Yijie/Xponge.git
cd Xponge
conda env create -f environment.yaml
conda activate <name-of-environment>
python setup.py install
```
