### 2022-08-05 Xia-Yijie/Xponge#3 @ojeda-e
[JOSS Review]: Package requires improved documentation
Description
-----------------
The documentation of the Xponge package is limited to the one provided in the README. Documentation of a scientific-based package usually requires doc pages in addition to README files.

Suggested solution
----------------------------
You can start by adding [docstrings](https://peps.python.org/pep-0257/) in your Python source code and implement [Sphinx](https://www.sphinx-doc.org/en/master/) to generate doc pages automatically. Additionally, for open-source packages, [ReadTheDocs](https://readthedocs.org/) host documentation online for free. Your package would benefit of using docstrings + Sphinx + ReadTheDocs.
Maybe you'll find this [tutorial](https://sphinx-tutorial.readthedocs.io/#sphinx-tutorial) helpful.

To fix this issue:
- [x]  add documentation pages. Docs should include installation instructions, listed dependencies, API, usage example, references.
- [x] reduce the length of your current README file and make it more readable for users that are not familiar with your code.
