### 2022-08-05 Xia-Yijie/Xponge#4 @ojeda-e
[JOSS review]: support across different OS
Description
-----------------
The README does not provide information about the supported platform. Given the lack of CI in the package is hard to judge the functionality on different Operating Systems (OS).

Question
-----------------
@Xia-Yijie, is the functionality of Xponge available in all SO (Linux/Mac/Windows)?

 This information should be provided to the user and should be clearly stated in the documentation

 To fix this issue:

- [x] Add supported OS to docs pages #3
