### 2022-08-08 Xia-Yijie/Xponge#5 @ojeda-e
[JOSS review]: Is the list of authors incomplete?
Description
----------------
The user with the GitHub handle @SummerJustNice seems to have made significant contributions to the Xponge package. (See [contributions link](https://github.com/Xia-Yijie/Xponge/graphs/contributors)).

Question
-------------
@Xia-Yijie , could you please provide a brief explanation on why @SummerJustNice is not included in list of authors? Is the list of authors incomplete?

Other suggestions:
--------------------------
If the list of authors is incomplete, please update your manuscript accordingly.
