### 2022-08-10 #4467 @hmacdope
@Xia-Yijie There are several serious initial acceptance blocker for mine on this.

I will raise an issue for each of these on the GitHub repository that @ojeda-e has not covered already.

1. A raw pip install on a clean conda environment fails to import correctly with `ModuleNotFoundError: No module named 'Xponge.load'`
2. Given this I could not run the tests and verify that the package works.
3. The base directory of the repository is cluttered with forcefield files organise these into a folder at a minimum. Additionally it is important to indicate whether these forcefields are distributed under any kind of license.
4. The base directory is cluttered with some files that look like junk eg [0.15_80_10_pH6.5_6lzg.result.pdb](https://github.com/Xia-Yijie/Xponge/blob/master/0.15_80_10_pH6.5_6lzg.result.pdb)
5. A large markdown file (README.md) is not sufficient documentation for a package of this complexity, you need to provide proper API docs.
6. You should provide some form of CI to prove that the package works. A simple github actions setup will do.
7. You should provide the citations of your dependent packages.
