### 2022-08-10 Xia-Yijie/Xponge#6 @hmacdope
[JOSS review] A pip install in a clean environment fails
A pip install of Xponge in a clean conda environment doesn't result in a functional import and fails with `ModuleNotFoundError: No module named 'Xponge.load'
`
