### 2022-08-10 Xia-Yijie/Xponge#7 @hmacdope
[JOSS review] Base directory cluttered with forcefield files that are not organised 
The base directory is cluttered with unorganised forcefield files and also has some junk?

- [x]  organise the FF files into some kind of folder and provide license terms for each and citations.
- [x] Remove or clarify role of  the apparent junk PDB files [0.15_80_10_pH6.5_6lzg.result.pdb](https://github.com/Xia-Yijie/Xponge/blob/master/0.15_80_10_pH6.5_6lzg.result.pdb) etc.
