### 2022-08-10 Xia-Yijie/Xponge#8 @hmacdope
[JOSS review] Provide some kind of CI to verify package works and is installable
The package needs some form of CI to verify that it works across different OS and that the package install works and tests pass.

Related to #4 #2 #6
A simple github actions setup will suffice.

- [x] Set up some CI
