### 2022-08-10 Xia-Yijie/Xponge#9 @hmacdope
[JOSS review] Cite packages used in README
If you use a package or encourage its use, please provide a citation.
