+++ lib/XpongeLib/parmchk_mod/common.c
@@ -1,1778 +1,0 @@
-#include <math.h>
-//#include <strings.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include "define.h"
-#include "eprintf.h"
-
-#define NUMBER_OF_CHEMICAL_ELEMENTS 118
-
-/*
-   Build the full path of an executable/lib/data file.
-   The executable/lib/file name is in the argument command.
-   The argument sizeof_path is the total size of the memory
-   pointed to by path, ie including space for the null terminator.
-   Return the size of the path it tried to build; thus, overflow
-   is indicated if the return value is greater than the sizeof_path;
-   However, on overflow an error is emitted and the program exits.
-   Assumptions:  strncat correctly handles a non-positive 3rd argument.
-   'for_system' indicates if the path will be used in a 'system'
-   command and therefore needs quotes if there are spaces in the
-   AMBERHOME variable.
-   dac change: code to handle spaces in AMBERHOME is wrong! remove earlier refs to quote
- */
-size_t build_path(char *path, const char *subdir, const char *fname, size_t sizeof_path,
-                  int for_system)
-{
-    /* number of chars copied plus 1; on overflow > sizeof_path */
-    size_t c = 1;
-    path[0] = '\0';
-
-    strncat(path, amberhome, sizeof_path - c);
-    c += strlen(amberhome);
-    strncat(path, subdir, sizeof_path - c);
-    c += strlen(subdir);
-    strncat(path, fname, sizeof_path - c);
-    c += strlen(fname);
-    if (c > sizeof_path) {
-        eprintf("Insufficient string size.\n"
-        "Increase MAXCHAR in define.h and rebuild.\n"
-        "Truncated string is \"%s\".", path);
-    }
-    return c;
-}
-
-
-size_t build_exe_path(char *path, const char *command, size_t sizeof_path, int for_system)
-{
-    return build_path(path, "/bin/", command, sizeof_path, for_system);
-}
-
-
-size_t build_dat_path(char *path, const char *command, size_t sizeof_path, int for_system)
-{
-    return build_path(path, "/dat/antechamber/", command, sizeof_path, for_system);
-}
-
-
-void default_minfo(MOLINFO * minfo)
-{
-    strcpy((*minfo).dkeyword, "CARTESIAN AM1 STANDARD DIRECT OPT=BFGS XTEST=0.0001");
-    strcpy((*minfo).mkeyword, "AM1 ANALYT MMOK GEO-OK PRECISE");
-
-/* RCW Below is the old setting from AMBER 12, 14. It is ridiculously tight. 
- * As of AmberTools 15 remove tight_p_conv and set grms_tol to 0.0005
- * tests show it makes no major difference to the charges - max diff = 0.001.
- *
- * strcpy((*minfo).skeyword, "  qm_theory='AM1', grms_tol=0.0002,\n  tight_p_conv=1, scfconv=1.d-10, ");
- *
-*/
-    strcpy((*minfo).skeyword,
-           "  qm_theory='AM1', grms_tol=0.0005,\n scfconv=1.d-10, ndiis_attempts=700, ");
-
-    strcpy((*minfo).gkeyword,
-           "#HF/6-31G* SCF=tight Test Pop=MK iop(6/33=2) iop(6/42=6) opt");
-    strcpy((*minfo).gesp, "g09.gesp");
-    strcpy((*minfo).ekeyword, "");
-    strcpy((*minfo).gm, "");
-    strcpy((*minfo).gn, "");
[…]
