+++ lib/XpongeLib/parmchk_mod/rotate.c
@@ -1,1272 +1,0 @@
-#include "eprintf.h"
-#define PAI 3.14159265359
-#define HALFPAI  1.570796326795
-
-
-double distance(ATOM at1, ATOM at2)
-{
-    double dist;
-    dist = (at1.x - at2.x) * (at1.x - at2.x);
-    dist += (at1.y - at2.y) * (at1.y - at2.y);
-    dist += (at1.z - at2.z) * (at1.z - at2.z);
-    dist = sqrt(dist);
-    return dist;
-}
-
-double rotate(ATOM obj1, ATOM obj2, ATOM obj3, ATOM * obj4)
-{
-
-/*
-This function is used to calculate the coordinate of the fourth atom in
-A-B-C-D system when we perform rotation bond B-C a certain degree
-The function can calculate the twist angle of four atoms before rotating
-and if you want twist angle be changed, please give you value(twist), the
-varible AngleChange should larger than 0
-*/
-
-    double twist = 0.0;
-    double w1, w2, w3, t;
-    double ya5, xc5, yd5, zd5;
-    double xa, ya, za, xc, yc, zc, xd, yd, zd;
-    double xa3, ya3, za3, xc3, zc3, xd3, yd3, zd3;
-    double ya4, za4, xc4, yd4, zd4;
-    /* double xd4 ; */
-
-    twist = twist / 180 * PAI;
-    xa = obj1.x - obj2.x;
-    ya = obj1.y - obj2.y;
-    za = obj1.z - obj2.z;
-    xc = obj3.x - obj2.x;
-    yc = obj3.y - obj2.y;
-    zc = obj3.z - obj2.z;
-    xd = (*obj4).x - obj2.x;
-    yd = (*obj4).y - obj2.y;
-    zd = (*obj4).z - obj2.z;
-    if (yc == 0)
-        w1 = 0.0;
-    else if (xc >= 0 && yc >= 0)
-        w1 = acos(xc / sqrt(xc * xc + yc * yc));
-    else if (xc <= 0 && yc >= 0)
-        w1 = -acos(-xc / sqrt(xc * xc + yc * yc)) + PAI;
-    else if (xc <= 0 && yc <= 0)
-        w1 = acos(-xc / sqrt(xc * xc + yc * yc)) - PAI;
-    else if (xc >= 0 && yc <= 0)
-        w1 = -acos(xc / sqrt(xc * xc + yc * yc));
-
-    /* rotating all the point clockwise about Z axis a w1 degree
-       then the atom C is lying in the Z-X plane */
-
-    xa3 = xa * cos(w1) + ya * sin(w1);
-    ya3 = -xa * sin(w1) + ya * cos(w1);
-    za3 = za;
-    xc3 = xc * cos(w1) + yc * sin(w1);
-
-    /*  yc3=-xc*sin(w1)+yc*cos(w1);
-       this value is not use anymore */
-
-    zc3 = zc;
-    xd3 = xd * cos(w1) + yd * sin(w1);
-    yd3 = -xd * sin(w1) + yd * cos(w1);
-    zd3 = zd;
-
-    /* rotating all  point clockwise about Y axis a w2 degree then
-       atom C is lying the X axis */
-
-    if (zc3 >= 0 && xc3 >= 0)
-        w2 = -acos(xc3 / sqrt(xc3 * xc3 + zc3 * zc3));
-    else if (zc3 <= 0 && xc3 >= 0)
-        w2 = acos(xc3 / sqrt(xc3 * xc3 + zc3 * zc3));
-    else if (zc3 <= 0 && xc3 <= 0)
[…]
