+++ lib/XpongeLib/__init__.py
@@ -1,3 +1,0 @@
-from . import backend
-
-_parmchk2 = backend._parmchk2
