+++ basic_requirements.yml
@@ -1,4 +1,0 @@
-name: Xponge
-dependencies:
-  - numpy
-
