### 2022-08-12 Xia-Yijie/Xponge#2 @Xia-Yijie
Thanks for your suggestions! I have set up some CI to check the installation, and also provide two `.yaml` files for the requirements.
