### 2022-08-12 Xia-Yijie/Xponge#7 @Xia-Yijie
I have organised the FF files and only keep them in the fold `Xponge/forcefield` now. I cite them in my `.py` file, and it can be seen after importing the module. For example, after importing `Xponge.forcefield.amber.ff14sb`, the citation for ff14SB can be seen. I can not find the license for the forcefields. I think the original authors may not have provided the license.
