### 2022-08-12 Xia-Yijie/Xponge#4 @Xia-Yijie
Yes, Xponge is available in all OS, but some functions to do the quantum chemistry calculations can not be used on Windows because `pyscf` is not available on Windows. I will write the information to the README.md and the detailed documentation.
