### 2022-08-12 Xia-Yijie/Xponge#1 @Xia-Yijie
I added an automated test by using `Xponge test`, but it was not implemented by the framework of Pytest or Unittest. Now I have re-implemented three tests by the framework Unittest, and I will write the instructions to the documentation.
