### 2022-08-12 e6c9d90
update README.md and paper.bib (for references)
