+++ paper.bib
@@ -229,3 +229,72 @@ year = {2011}
     title = {LigParGen Server},
     url = {http://zarbi.chem.yale.edu/ligpargen/}  
 }
+
+@misc{rdkit,
+    author = {Greg Landrum, Paolo Tosco, Brian Kelley, Ric, sriniker, gedeck, Riccardo Vianello, NadineSchneider, Eisuke Kawashima, David Cosgrove, Andrew Dalke, Dan N, Gareth Jones, Brian Cole, Matt Swain, Samo Turk, AlexanderSavelyev, Alain Vaucher, Maciej Wójcikowski, Ichiru Take, Daniel Probst, Kazuya Ujihara, Vincent F. Scalfani, guillaume godin, Axel Pahl, Francois Berenger, JLVarjo, strets123, JP, DoliathGavid},
+    year = {2022},
+    title = {rdkit/rdkit: 2022_03_5 (Q1 2022) Release},
+    url = {https://zenodo.org/record/6961488#.YvackPhByUk}  
+}
+
+@misc{mindspore,
+    author = {Mindspore developer},
+    year = {2022},
+    title = {MindSpore: An Open AI Framwork},
+    url = {http://www.mindspore.cn}  
+}
+
+@article{pyscf,
+author = {Sun,Qiming  and Zhang,Xing  and Banerjee,Samragni  and Bao,Peng  and Barbry,Marc  and Blunt,Nick S.  and Bogdanov,Nikolay A.  and Booth,George H.  and Chen,Jia  and Cui,Zhi-Hao  and Eriksen,Janus J.  and Gao,Yang  and Guo,Sheng  and Hermann,Jan  and Hermes,Matthew R.  and Koh,Kevin  and Koval,Peter  and Lehtola,Susi  and Li,Zhendong  and Liu,Junzi  and Mardirossian,Narbe  and McClain,James D.  and Motta,Mario  and Mussard,Bastien  and Pham,Hung Q.  and Pulkin,Artem  and Purwanto,Wirawan  and Robinson,Paul J.  and Ronca,Enrico  and Sayfutyarova,Elvira R.  and Scheurer,Maximilian  and Schurkus,Henry F.  and Smith,James E. T.  and Sun,Chong  and Sun,Shi-Ning  and Upadhyay,Shiv  and Wagner,Lucas K.  and Wang,Xiao  and White,Alec  and Whitfield,James Daniel  and Williamson,Mark J.  and Wouters,Sebastian  and Yang,Jun  and Yu,Jason M.  and Zhu,Tianyu  and Berkelbach,Timothy C.  and Sharma,Sandeep  and Sokolov,Alexander Yu.  and Chan,Garnet Kin-Lic },
+title = {Recent developments in the PySCF program package},
+journal = {The Journal of Chemical Physics},
+volume = {153},
+number = {2},
+pages = {024109},
+year = {2020},
+doi = {10.1063/5.0006074},
+URL = {https://doi.org/10.1063/5.0006074},
+eprint = {https://doi.org/10.1063/5.0006074}
+}
+
+@article{pyscf2,
+author = {Sun, Qiming and Berkelbach, Timothy C. and Blunt, Nick S. and Booth, George H. and Guo, Sheng and Li, Zhendong and Liu, Junzi and McClain, James D. and Sayfutyarova, Elvira R. and Sharma, Sandeep and Wouters, Sebastian and Chan, Garnet Kin-Lic},
+title = {PySCF: the Python-based simulations of chemistry framework},
+journal = {WIREs Computational Molecular Science},
+volume = {8},
+number = {1},
+pages = {e1340},
+doi = {https://doi.org/10.1002/wcms.1340},
[…]
