+++ README.md
@@ -1,16 +1,16 @@
 # Welcome to Use Xponge!
 
-## What is Xponge?
+## Introduction
 
 ``Xponge`` is a lightweight and easy-customizing python package to perform pre- and post-processing of molecular simulations.
 
-## What can Xponge do?
+### What can Xponge do?
 
-Xponge includes three major categories of functionality, namely, the simulation system construction, simulation data transformation and analysis, and automated workflows for complex simulations. ``Xponge`` is mainly designed for the molecular dynamics (MD) program [SPONGE](https://onlinelibrary.wiley.com/doi/epdf/10.1002/cjoc.202100456), but it can also output some general format files such as mol2 and PDB, so it may help the other molecular modelling programs too.
+Xponge includes three major categories of functionality, namely, the simulation system construction, simulation data transformation and analysis, and automated workflows for complex simulations. ``Xponge`` is mainly designed for the molecular dynamics (MD) program [SPONGE](https://onlinelibrary.wiley.com/doi/epdf/10.1002/cjoc.202100456)[1], but it can also output some general format files such as mol2 and PDB, so it may help the other molecular modelling programs too.
 
-## How can I get Xponge?
+## Installation
 
-Xponge can be used on both Windows and Linux operating systems
+Xponge can be used on all operating systems (Windows/Linux/MacOS).
 
 ### 1. pip install
 
@@ -25,6 +25,9 @@ pip install Xponge
     The gitee repository is [here](https://gitee.com/gao_hyp_xyj_admin/xponge).
     The github repository is [here](https://github.com/xia-yijie/xponge).
 
+        git clone http://gitee.com/gao_hyp_xyj_admin/xponge.git
+        git clone http://github.com/xia-yijie/xponge.git
+
 - 2.2 open the directory where you download or clone the repository
 
 - 2.3 run the command
@@ -33,32 +36,75 @@ pip install Xponge
     python setup.py install
     ```
 
-## How can I check whether I have installed Xponge correctly?
+### Installation check
 
 There are some unit tests in ``Xponge``. You can do the basic test to check whether the installation is successful like this:
 
 ```bash
-Xponge test -do base -o test
+Xponge test --do base -o test --verbose 1
 ```
 
-Here, ``Xponge`` can be replaced to ``python -m Xponge``, ``python3 -m Xponge`` and so on according to your settings of the environmental variables.
+Here, ``Xponge`` can be replaced to ``python -m Xponge``, ``python3 -m Xponge`` and so on according to your settings of the environmental variables. Some files will be generated after the test is finished.
 
-## Can you give me a example to use?
+## Usage example
 
 Here is a simple example.
 
 ```python
 import Xponge
-# import the force field you need
+# Import the force field you need
 import Xponge.forcefield.amber.ff14sb
-# build the molecule
+# Build the molecule like this
 peptide = ACE + ALA + NME
-# save it as your favorite format
+# or like this
+peptide2 = NALA + ALA * 10 + CALA
+# or like this
+peptide3 = Xponge.Get_Peptide_From_Sequence("AAAAA")
+# See the documentation for more usage!
+# Save them as your favorite format
 Xponge.save_pdb(peptide, "ala.pdb")
-Xponge.save_mol2(peptide, "ala.mol2")
-Xponge.save_sponge_input(peptide, "ala")
+Xponge.save_mol2(peptide2, "ala12.mol2")
+Xponge.save_sponge_input(peptide3, "ala5")
 ```
 
[…]
