+++ Xponge/mdrun/__main__.py
@@ -4,5 +4,8 @@ This **module** helps to run the molecular dynamics simulation
 import sys
 from . import run
 
+def main():
+    run(sys.argv)
 
-run(sys.argv)
+if __name__ == "__main__":
+    main()
