+++ docs/source/automated_workflow.rst
@@ -1,0 +1,5 @@
+automated workflow
+----------------------
+.. include:: namespace.rst
+
+See `relative binding energy of FXR <https://gitee.com/mindspore/mindscience/blob/master/MindSPONGE/applications/molecular_dynamics/tradition/relative_binding_energy_of_FXR/relative_binding_energy_of_FXR.ipynb>`_ for example.
