+++ docs/source/usage21.rst
@@ -1,23 +1,0 @@
-developing forcefield
-----------------------
-.. include:: namespace.rst
-
-new parameters
-====================
-
-
-new forcefield format
-=======================
-
-non-bonded interactions
-##########################
-
-
-bonded interactions
-##########################
-
-new forcefield combination
-============================
-
-new type assignment
-=======================
