+++ docs/source/usage13.rst
@@ -1,5 +1,0 @@
-structure pre-process
----------------------
-.. include:: namespace.rst
-
-todo
