+++ docs/source/usage15.rst
@@ -1,5 +1,0 @@
-command line tools
---------------------
-.. include:: namespace.rst
-
-todo

[…]
