### 2022-08-16 3a7e255
make load_pdb more compatible
