+++ Xponge/__init__.py
@@ -89,7 +89,7 @@ from .helper import GlobalSetting, Type, ResidueType, Entity, Atom, Residue, Res
 from .load import load_ffitp, load_mol2, load_rst7, load_frcmod, load_pdb, load_parmdat, load_coordinate
 from .build import save_mol2, save_pdb, save_sponge_input, save_gro, build_bonded_force, get_mindsponge_system_energy
 from .process import impose_bond, impose_angle, impose_dihedral, add_solvent_box, h_mass_repartition, solvent_replace, \
-    main_axis_rotate, get_peptide_from_sequence
+    main_axis_rotate, get_peptide_from_sequence, optimize
 
 
 def _initialize():
