### 2022-08-17 4b68af1
new functions: lammps-like lattice building system
