+++ README.md
@@ -83,8 +83,8 @@ region_2 = BlockRegion(0, 0, 40, 20, 20, 60)
 region_3 = BlockRegion(0, 0, 0, 20, 20, 20)
 region_4 = SphereRegion(20, 10, 30, 10)
 region_5 = BlockRegion(20, 0, 20, 60, 60, 60)
-region_2or3 = Region(region_2, region_3, do="union")
-region_4and5 = Region(region_4, region_5, do="intersect")
+region_2or3 = UnionRegion(region_2, region_3)
+region_4and5 = IntersectRegion(region_4, region_5)
 t = Lattice("bcc", basis_molecule=CL, scale=4)
 t2 = Lattice("fcc", basis_molecule=K, scale=3)
 t3 = Lattice("sc", basis_molecule=NA, scale=3)
