### 2022-08-20 2635567
add the common class AbstractMolecule to reuse code and make code clean and dry
