+++ Xponge/helper/__init__.py
@@ -9,9 +9,10 @@ import os
 import time
 import stat
 from types import MethodType, FunctionType
-from functools import partial, wraps
+from functools import partial, partialmethod, wraps
 from collections import OrderedDict
 from itertools import product
+from abc import ABC
 
 import numpy as np
 
@@ -199,7 +200,7 @@ class _GlobalSetting():
 globals()["GlobalSetting"] = _GlobalSetting()
 
 
-class Type:
+class Type(ABC):
     """
     This **class** is the abstract class of the types (atom types, bonded force types and so on).
 
@@ -483,6 +484,10 @@ class Type:
         type(self)._unit_transfer(self)
 
 
+class AbstractMolecule(ABC):
+    """This abstract **class** is used to judge whether a class can be treated as a molecule """
+
+
 class AtomType(Type):
     """
     This **class** is a subclass of Type, for atom types
@@ -754,7 +759,7 @@ class ResidueType(Type):
         return new_restype
 
 
-class Entity:
+class Entity(ABC):
     """
     This **class** is the abstract class of the entities (atoms, bonded forces, residues and so on).
 
@@ -1114,13 +1119,35 @@ class ResidueLink:
         """indicates the instance built or not"""
         self.bonded_forces = {frc.get_class_name(): [] for frc in GlobalSetting.BondedForces}
         """the bonded forces after building"""
+        self.tohash = self.get_hash(atom1, atom2)
+        self.residue_tohash = self.get_hash(atom1, atom2, True)
         set_attribute_alternative_names(self)
 
     def __repr__(self):
         return "Entity of ResidueLink: " + repr(self.atom1) + "-" + repr(self.atom2)
 
     def __hash__(self):
-        return hash(repr(self))
+        return hash(self.tohash)
+
+    def __eq__(self, other):
+        return isinstance(other, ResidueLink) and self.tohash == other.tohash
+
+    @staticmethod
+    def get_hash(atom1, atom2, residue=False):
+        """
+        This **function** is used to get the hash value of the ResidueLink between atom0 and atom1
+
+        :param atom1: the first atom to link
+        :param atom2: the second atom to link
+        :param residue: whether to hash the residue of the atoms or atoms themselves
+        :return: the hash value
+        """
+        if residue:
+            tohash = [f"{atom1}-{atom2}", f"{atom2}-{atom1}"]
+        else:
+            tohash = [f"{atom1.residue}-{atom2.residue}", f"{atom2.residue}-{atom1.residue}"]
+        tohash.sort()
+        return tohash[0]
 
     def add_bonded_force(self, bonded_force_entity):
         """
@@ -1145,7 +1172,7 @@ class ResidueLink:
         raise Exception("Not Right forcopy")
 
 
-class Molecule:
+class Molecule():
     """
     This **class** is a class for molecules
 
@@ -1175,7 +1202,7 @@ class Molecule:
         """
         self.atom_index = []
         """the atom index"""
-        self.residue_links = []
+        self.residue_links = set()
         """the residue links in the molecule"""
         self.bonded_forces = Xdict()
         """the bonded forces after building"""
@@ -1200,6 +1227,34 @@ class Molecule:
         if getattr(AtomType, "_parameters").get(attr, None) == float:
             return np.sum([getattr(atom, attr) for res in self.residues for atom in res.atoms])
         return super().__getattribute__(attr)
+
+    @classmethod
+    def cast(cls, other, deepcopy=False):
+        """
+        This **function** casts a Residue, a ResidueType or a Molecule to a Molecule
+
[…]
