+++ Xponge/tools/__init__.py
@@ -165,6 +165,23 @@ class TestMyPackage(unittest.TestCase):
         mol = t5.create(box, region_7, mol)
         Save_PDB(mol, f"{args.o}.pdb")
 
+    def test_fep(self):
+        """
+        This **function** does the assignment test for the FEP functions
+
+        :param args: arguments from argparse
+        :return: None
+        """
+        source("..")
+        source("..forcefield.amber.ff14sb")
+        source("__main__")
+        args = self.args
+        Save_PDB(ALA,f"{args.o}.pdb")
+        Save_Mol2(ALA, f"{args.o}_r1.mol2")
+        Save_Mol2(NALA,f"{args.o}_r2.mol2")
+        error = os.system(f"Xponge mol2rfe -nl 1 -pdb {args.o}.pdb -r1 {args.o}_r1.mol2 \
+-r2 {args.o}_r2.mol2 > {os.devnull}")
+        self.assertEqual(error, 0)
 
 def _one_test(ccon, name, args):
     """
@@ -223,7 +240,7 @@ def test(args):
         Xprint("No error or failure", verbose=-1)
     else:
         Xprint(f"{len(errors)} error(s) and {len(failures)} failure(s) found", verbose=-1)
-        sys.exit(1)
+        sys.exit(len(errors) + len(failures))
 
 
 def converter(args):
