### 2022-08-21 #4467 @hmacdope
@Xia-Yijie some things to address in the paper.

- [ ]  "Easy customizing" should be "Easy to customize"
- [ ]  "tip3p, tip4pew, opc for water" The names of the water models should be in all caps and need citations.
- [ ]  What is meant by "simulation data transformation", please clarify
- [ ]  "hyd ration" has an extra space and should be "hydration"
- [ ]  "small organic molecules in  water is now widely used in drug design" This sentence needs a citation.
- [ ]  I would add the ATB https://atb.uq.edu.au/ to the list of online parametrisation engines.
- [ ]  "and can be difficult to be customized" should be "can be difficult to customize"
- [ ]  "Meanwhile, since machine learning force field for molecular simulation represents one of the current research frontiers"  This sentence needs a citation.
- [ ] "Xponge is written by python and thus can be better merged into the computational graph." The meaning of this is unclear. What is meant by computational graph in this context.
- [ ] Please provide some examples where XPonge has been used in published work if applicable
- [ ] Could you clarify how Xponge is modular, thereby providing an advantage over existing tools?
- [ ] Could you clarify the role of the XpongeLib package in Xponge?
