### 2022-08-22 8303896
Update docstrings in __init__.py
