+++ paper.bib
@@ -237,8 +237,8 @@ year = {2011}
     url = {https://zenodo.org/record/6961488#.YvackPhByUk}  
 }
 
-@misc{mindspore,
-    author = {Mindspore developer},
+@misc{mindSpore,
+    author = {MindSpore developer},
     year = {2022},
     title = {MindSpore: An Open AI Framwork},
     url = {http://www.mindspore.cn}  
@@ -297,4 +297,118 @@ year = {2016},
 doi = {10.1063/1.4952956},
 URL = {https://doi.org/10.1063/1.4952956},
 eprint = {https://doi.org/10.1063/1.4952956}
+}
+
+@article{Jorgensen:1983,
+author = {Jorgensen,William L.  and Chandrasekhar,Jayaraman  and Madura,Jeffry D.  and Impey,Roger W.  and Klein,Michael L. },
+title = {Comparison of simple potential functions for simulating liquid water},
+journal = {The Journal of Chemical Physics},
+volume = {79},
+number = {2},
+pages = {926-935},
+year = {1983},
+doi = {10.1063/1.445869},
+URL = {https://doi.org/10.1063/1.445869},
+eprint = {https://doi.org/10.1063/1.445869}
+}
+
+@article{Horn:2004,
+author = {Horn,Hans W.  and Swope,William C.  and Pitera,Jed W.  and Madura,Jeffry D.  and Dick,Thomas J.  and Hura,Greg L.  and Head-Gordon,Teresa },
+title = {Development of an improved four-site water model for biomolecular simulations: TIP4P-Ew},
+journal = {The Journal of Chemical Physics},
+volume = {120},
+number = {20},
+pages = {9665-9678},
+year = {2004},
+doi = {10.1063/1.1683075},
+URL = {https://doi.org/10.1063/1.1683075},
+eprint = {https://doi.org/10.1063/1.1683075}
+}
+
+@article{Izadi:2014,
+author = {Izadi, Saeed and Anandakrishnan, Ramu and Onufriev, Alexey V.},
+title = {Building Water Models: A Different Approach},
+journal = {The Journal of Physical Chemistry Letters},
+volume = {5},
+number = {21},
+pages = {3863-3871},
+year = {2014},
+doi = {10.1021/jz501780a},
+note ={PMID: 25400877},
+URL = {https://doi.org/10.1021/jz501780a},
+eprint = {https://doi.org/10.1021/jz501780a}
+}
+
+@inbook{Yang:2016,
+author = {Yang, L. and Zhang, Jun and Che, Xing and Gao, Yi},
+year = {2016},
+month = {06},
+pages = {},
+title = {Simulation Studies of Protein and Small Molecule Interactions and Reaction},
+volume = {578},
+isbn = {9780128111079},
+journal = {Methods in Enzymology},
+doi = {10.1016/bs.mie.2016.05.031}
+}
+
+@article{Malde:2011,
+author = {Malde, Alpeshkumar K. and Zuo, Le and Breeze, Matthew and Stroet, Martin and Poger, David and Nair, Pramod C. and Oostenbrink, Chris and Mark, Alan E.},
+title = {An Automated Force Field Topology Builder (ATB) and Repository: Version 1.0},
+journal = {Journal of Chemical Theory and Computation},
+volume = {7},
+number = {12},
+pages = {4026-4037},
+year = {2011},
+doi = {10.1021/ct200196m},
+    note ={PMID: 26598349},
+URL = {https://doi.org/10.1021/ct200196m},
+eprint = {https://doi.org/10.1021/ct200196m}
+}
+
+@article{Unke:2021,
[…]
