+++ paper.bib
@@ -412,3 +412,10 @@ eprint = {https://doi.org/10.1021/j100308a038}
     title = {The State of the Octoverse: machine learning},
     url = {https://github.blog/2019-01-24-the-state-of-the-octoverse-machine-learning/}  
 }
+
+@misc{ComputationalGraph,
+    author = { Tutorials Point (I) Pvt. Ltd.},
+    year = {2018},
+    title = {Python Deep Learning: Computational Graphs},
+    url = {https://www.tutorialspoint.com/python_deep_learning/python_deep_learning_computational_graphs.htm}  
+}
