+++ contribution_guide.rst
@@ -1,0 +1,54 @@
+Contribution Guide
+=====================
+
+The repository of Xponge is both in `Gitee <https://gitee.com/gao_hyp_xyj_admin/xponge>`_ and `Github <https://github.com/Xia-Yijie/xponge>`_ under the Apache 2.0 license. 
+
+You can contribute to Xponge in both Gitee or Github, but Github is not directly accessible in Chine due to DNS cache poisoning, so it is recommended to give the suggestions or create the pull request to the Gitee repository.
+
+Bug Reports or Enhancement Requests
+######################################
+
+You can use the issue tracking system of the Gitee or Github to report the bug you meet or request the enhancement you need.
+
+It is okay to raise an issue either in English or Chinese.
+
+1. gitee
+
+    .. image:: https://gitee.com/gao_hyp_xyj_admin/xponge/raw/master/README_PICTURE/3.jpg
+    .. image:: https://gitee.com/gao_hyp_xyj_admin/xponge/raw/master/README_PICTURE/4.jpg
+
+2. github
+
+    .. image:: https://gitee.com/gao_hyp_xyj_admin/xponge/raw/master/README_PICTURE/5.jpg
+    .. image:: https://gitee.com/gao_hyp_xyj_admin/xponge/raw/master/README_PICTURE/6.jpg
+
+Contributing to the main codebase
+#####################################
+
+You can use the pull request system of the Gitee or Github to contribute to the main codebase.
+
+Please do not use East Asian characters in the code to avoid confusion in different operating systems.
+
+1. First of all, you need to fork the repository of the Gitee or Github to report
+
+    - gitee
+
+        .. image:: https://gitee.com/gao_hyp_xyj_admin/xponge/raw/master/README_PICTURE/7.jpg
+
+    - github
+
+        .. image:: https://gitee.com/gao_hyp_xyj_admin/xponge/raw/master/README_PICTURE/8.jpg
+
+
+2. Write your code to your repository
+
+3. Create a pull request
+
+    - gitee
+
+        .. image:: https://gitee.com/gao_hyp_xyj_admin/xponge/raw/master/README_PICTURE/9.jpg
+
+    - github
+
+        .. image:: https://gitee.com/gao_hyp_xyj_admin/xponge/raw/master/README_PICTURE/15.png
+
