+++ paper.bib
@@ -238,7 +238,7 @@ year = {2011}
 }
 
 @misc{mindSpore,
-    author = {MindSpore developer},
+    author = {MindSpore},
     year = {2022},
     title = {MindSpore: An Open AI Framwork},
     url = {http://www.mindspore.cn}  
@@ -414,7 +414,7 @@ eprint = {https://doi.org/10.1021/j100308a038}
 }
 
 @misc{ComputationalGraph,
-    author = { Tutorials Point (I) Pvt. Ltd.},
+    author = {TutorialsPoint},
     year = {2018},
     title = {Python Deep Learning: Computational Graphs},
     url = {https://www.tutorialspoint.com/python_deep_learning/python_deep_learning_computational_graphs.htm}  
