+++ paper.md
@@ -26,7 +26,7 @@ Xponge is a lightweight and easy to customize Python package to perform pre- and
 # Statement of need
 
 Molecular simulation is becoming an important and useful tool in many different research areas. For example, the computational simulation of proteins and small organic molecules in water is now widely used in drug design [@Yang:2016]. In MD simulations, it is important to construct and parameterize the simulation system and to process and analyze the simulation results. Because of the complexity and variety of systems to be simulated, it is often necessary to develop new and different force fields, approximations, algorithms, or analytical methods to obtain and visualize results.
-There exist a variety of pre-processing tools for molecular simulations, such as offline LEaP [@LEaP], pdb2gmx [@pdb2gmx] and psfgen [@psfgen], and online LigParGen [@LigParGen], CGenFF [@CGenFF] and ATB [@Malde:2011], but they are designed for a particular form of force field and can be difficult to customize. Xponge is designed to be highly modular and intended to be developer-friendly. When you want to develop your own force field, It is hard to modify the online tools and the offline tools usually only have parameter Interfaces. While Xponge is like building blocks, developers can write their own force field modules outside Xponge to implement the functionality they want. At the same time, Xponge includes a number of post-processing capabilities and combines specific complex pre-processing - simulation - post-processing into one single workflow. Meanwhile, since machine learning force field for molecular simulation represents one of the current research frontiers [@Unke:2021; Poltavsky:2021], and Python is now the top programming language for machine learning [@Elliott:2019]. Xponge is written by python and has interfaces to call the machine learning framework MindSpore [@MindSpore], thus Xponge can be better merged into the computational graph [@ComputationalGraph] of machine learning. Such a setup fits well with molecular simulation softwares such as SPONGE [@Huang:2022] which incorporates machine learning methods.
+There exist a variety of pre-processing tools for molecular simulations, such as offline LEaP [@LEaP], pdb2gmx [@pdb2gmx] and psfgen [@psfgen], and online LigParGen [@LigParGen], CGenFF [@CGenFF] and ATB [@Malde:2011], but they are designed for a particular form of force field and can be difficult to customize. Xponge is designed to be highly modular and intended to be developer-friendly. When you want to develop your own force field, It is hard to modify the online tools and the offline tools usually only have parameter interfaces. Xponge is like building blocks, developers can write their own force field modules (even outside Xponge) to implement the functionality they want. At the same time, Xponge includes a number of post-processing capabilities and combines specific complex pre-processing - simulation - post-processing into one single workflow. Meanwhile, since machine learning force field for molecular simulation represents one of the current research frontiers [@Unke:2021; @Poltavsky:2021], and Python is now the top programming language for machine learning [@Elliott:2019]. Xponge is written by python and has interfaces to call the machine learning framework MindSpore [@MindSpore], thus Xponge can be better merged into the computational graph [@ComputationalGraph] of machine learning. Such a setup fits well with molecular simulation softwares such as SPONGE [@Huang:2022] which incorporates machine learning methods.
 
 # Availability
 
