### 2022-08-22 #4467 @Xia-Yijie
> "Easy customizing" should be "Easy to customize"

I have corrected the expression in the paper, the README file and the documentation.

> "tip3p, tip4pew, opc for water" The names of the water models should be in all caps and need citations.

I have corrected the names and added the citations.

>  What is meant by "simulation data transformation", please clarify

I have added a sentence "Simulation data transformation is the process of changing the format, structure, or values of the simulation data such as molecular dynamics trajectories and coordinates from the final frame of a simulation." to the paper.

> "hyd ration" has an extra space and should be "hydration"

I have corrected the spelling mistake.

> "small organic molecules in water is now widely used in drug design" This sentence needs a citation.

I have added a citation. By the way, this sentence should be "proteins and small organic molecules in water is now widely used in drug design".

> I would add the ATB https://atb.uq.edu.au/ to the list of online parametrisation engines.

I have added the ATB to the list, and cited the description and the latest validation of the publications.

> "and can be difficult to be customized" should be "can be difficult to customize"

I have corrected the expression.

>  "Meanwhile, since machine learning force field for molecular simulation represents one of the current research frontiers" This sentence needs a citation.

I have added two citations.

>  "Xponge is written by python and thus can be better merged into the computational graph." The meaning of this is unclear. What is meant by computational graph in this context.

The word "computational graph" is a term in machine learning, and machine learning is mentioned in the context. I think the explanation of the word "computational graph" is out of the main idea of the paper, so I just give a citation in TutorialsPoint here.

> Please provide some examples where Xponge has been used in published work if applicable

There is no published work which has used Xponge now. Some articles using Xponge are still under peer review.

> Could you clarify how Xponge is modular, thereby providing an advantage over existing tools?

I have added the following sentences to the paper.

"When a developer wants to develop his own force field, it is hard to modify the online tools and the offline tools usually only have parameter interfaces. While in Xponge, the core only defines the basic classes. No function is implemented, and only the interfaces are left. The force field modules use the interfaces to implement the functions they need themselves. In fact, the implementation of force fields is more like scripts than modules. Developers can write their own force fields (even outside Xponge) to implement the features they want."

>  Could you clarify the role of the XpongeLib package in Xponge?

Xponge uses look-up tables to parameterize the force field. In some general force fields (UFF, GAFF  for example), there are a large number of atom types, and it is possible to form a bond between any two atom types. As a result, there are too many permutations of atomic types to generate a look-up table for all of the bonded forces (O(N^4) for dihedral, for example), so a C/C++ library is used to generate the look-up table which only contains the atom types used in the working system.
