### 2022-08-23 #4467 @hmacdope
Just a  few more queries,

> I have added a sentence "Simulation data transformation is the process of changing the format, structure, or values of the simulation data such as molecular dynamics trajectories and coordinates from the final frame of a simulation." to the paper.

It doesn't have to be the final frame of a simulation, it can be any frame no?

> "When a developer wants to develop his own force field, it is hard to modify the online tools and the offline tools usually only have parameter interfaces. While in Xponge, the core only defines the basic classes. No function is implemented, and only the interfaces are left. The force field modules use the interfaces to implement the functions they need themselves. In fact, the implementation of force fields is more like scripts than modules. Developers can write their own force fields (even outside Xponge) to implement the features they want."

This paragraph was not clear to me as it is unclear what is meant by "basic classes" and "no function is implemented". Perhaps something along the lines of. " Xponge provides an interface that allows users to specify their forcefields themselves"
