### 2022-08-23 #4467 @hmacdope
@Xia-Yijie please fix the invalid DOIs as well.
