+++ paper.md
@@ -21,12 +21,12 @@ bibliography: paper.bib
 
 # Summary
 
-Xponge is a lightweight and easy to customize Python package to perform pre- and post-processing of molecular simulations. It is mainly designed for the MD program SPONGE [@Huang:2022], but it can also process common format files and therefore it should also be useful for other simulation packages such as GROMACS [@Abraham:2015] and LAMMPS [@Thompson:2022]. Xponge includes three major categories of functionality, namely, the simulation system construction, simulation data transformation and analysis, and automated workflows for complex simulations. For the construction of simulation systems, Xponge can generate 3-dimensional molecular structures or read structures downloaded from online databases such as RCSB [@Berman:2000] and PubChem [@Sayers:2021], and perform force field parameterization. The current force fields supported by Xponge include CHARMM27 [@MacKerell:1998; @Mackerell:2004], ff14SB [@Maier:2015], ff19SB [@Tian:2020] for proteins, lipid14 [@Dickson:2014] and lipid17 for lipids, GAFF [@Wang:2004] for small organic molecules and SPC/E [@Berendsen:1987], TIP3P [@Jorgensen:1983], TIP4P-Ew [@Horn:2004], OPC [@Izadi:2014] for water. Simulation data transformation is the process of changing the format, structure, or values of the simulation data such as molecular dynamics trajectories and coordinates from the final frame of a simulation. For both simulation data transformation and analysis, Xponge is combined with the Python package MDAnalysis [@Gowers:2016; @Michaud-Agrawal:2011] for processing. Xponge has an integrated workflow for free energy perturbation calculations with dual topology construction, facilitating calculations on hydration and binding free energies. At the same time, Xponge itself is highly modular and easily customizable, enabling simple extensions that mimic existing modules to develop one's own force fields, data analytics, and workflows.
+Xponge is a lightweight and easy to customize Python package to perform pre- and post-processing of molecular simulations. It is mainly designed for the MD program SPONGE [@Huang:2022], but it can also process common format files and therefore it should also be useful for other simulation packages such as GROMACS [@Abraham:2015] and LAMMPS [@Thompson:2022]. Xponge includes three major categories of functionality, namely, the simulation system construction, simulation data transformation and analysis, and automated workflows for complex simulations. For the construction of simulation systems, Xponge can generate 3-dimensional molecular structures or read structures downloaded from online databases such as RCSB [@Berman:2000] and PubChem [@Sayers:2021], and perform force field parameterization. The current force fields supported by Xponge include CHARMM27 [@MacKerell:1998; @Mackerell:2004], ff14SB [@Maier:2015], ff19SB [@Tian:2020] for proteins, lipid14 [@Dickson:2014] and lipid17 for lipids, GAFF [@Wang:2004] for small organic molecules and SPC/E [@Berendsen:1987], TIP3P [@Jorgensen:1983], TIP4P-Ew [@Horn:2004], OPC [@Izadi:2014] for water. Simulation data transformation is the process of changing the format, structure, or values of the simulation data such as molecular dynamics trajectories and coordinates from any frame of a simulation. For both simulation data transformation and analysis, Xponge is combined with the Python package MDAnalysis [@Gowers:2016; @Michaud-Agrawal:2011] for processing. Xponge has an integrated workflow for free energy perturbation calculations with dual topology construction, facilitating calculations on hydration and binding free energies. At the same time, Xponge itself is highly modular and easily customizable, enabling simple extensions that mimic existing modules to develop one's own force fields, data analytics, and workflows.
 
 # Statement of need
 
[…]
