### 2022-08-27 Xia-Yijie/Xponge#2 @ojeda-e
Thanks.
I noticed that you added two yaml files but couldn't find where in docs or the readme you provide instructions on how to install Xponge using them. I also couldn't find information on what is the difference between both.
If the information above is missing, please add it to docs with a link in the readme. Otherwise please share the link.
