+++ README.md
@@ -103,9 +103,9 @@ region_4 = Xponge.SphereRegion(20, 10, 30, 10)
 region_5 = Xponge.BlockRegion(20, 0, 20, 60, 60, 60)
 region_2or3 = Xponge.UnionRegion(region_2, region_3)
 region_4and5 = Xponge.IntersectRegion(region_4, region_5)
-t = Lattice("bcc", basis_molecule=CL, scale=4)
-t2 = Lattice("fcc", basis_molecule=K, scale=3)
-t3 = Lattice("sc", basis_molecule=NA, scale=3)
+t = Xponge.Lattice("bcc", basis_molecule=CL, scale=4)
+t2 = Xponge.Lattice("fcc", basis_molecule=K, scale=3)
+t3 = Xponge.Lattice("sc", basis_molecule=NA, scale=3)
 mol = t.Create(box, region_1)
 mol = t2.create(box, region_2or3, mol)
 mol = t3.create(box, region_4and5, mol)
