### 2022-08-28 Xia-Yijie/Xponge#2 @Xia-Yijie
The information was missing, and now I have added the information to [docs](https://spongemm.cn/xponge_doc/installation.html#id2) and the [readme](https://github.com/Xia-Yijie/Xponge#2-source-setup).
