### 2022-08-30 Xia-Yijie/Xponge#2 @ojeda-e
Just one more thing to close this issue. Would you please clarify what is the difference between the basic functionality and full functionality? How would the user know which env to install?
