### 2022-09-05 Xia-Yijie/Xponge#2 @Xia-Yijie
@ojeda-e I have changed the wording of the section in the [docs](https://spongemm.cn/xponge_doc/installation.html#id2) and the [readme](https://github.com/Xia-Yijie/Xponge#2-source-setup), do you think it is enough now to make users understand the difference between the two files?
