### 2022-09-05 Xia-Yijie/Xponge#4 @Xia-Yijie
@ojeda-e I have added the information to the [README](https://github.com/Xia-Yijie/Xponge/blob/master/README.md#installation) (the first paragraph of installation) and [docs](https://spongemm.cn/xponge_doc/index.html) (the section `How can I get Xponge` in the index page).
