### 2022-09-05 5ecd566
update the interface with mindspore 1.8 (#21)
* update docs

* update the interface for the newest updated MindSPONGE
