+++ Xponge/build.py
@@ -3,6 +3,7 @@ This **module** is used to build and save
 """
 import os
 from itertools import product
+from warnings import warn
 from . import assign
 from .helper import AbstractMolecule, ResidueType, Molecule, Residue, ResidueLink, GlobalSetting, Xopen, Xdict, \
     set_global_alternative_names
@@ -343,9 +344,8 @@ def get_mindsponge_system_energy(cls, use_pbc=False):
     from mindsponge import set_global_units
     from mindsponge import Molecule as mMolecule
     from mindsponge import ForceFieldBase
-    #pylint: disable=deprecated-class
-    from collections import Sequence
-    if not isinstance(cls, Sequence):
+    from collections.abc import Iterable
+    if not isinstance(cls, Iterable):
         cls = [cls]
     sys_kwarg = Xdict()
     ene_kwarg = Xdict()
@@ -417,6 +417,132 @@ def save_sponge_input(cls, prefix=None, dirname="."):
         save_sponge_input(residue, prefix, dirname)
 
 
+def _pdb_chain(cls: Molecule):
+    """
+
+    :param cls:
+    :return:
+    """
+    chains = Xdict()
+    chain_ids = [" "] * len(cls.residues)
+    start = 0
+    alphabet = "A"
+    i = 0
+    for i, res in enumerate(cls.residues):
+        if i == start:
+            continue
+        if not cls.get_residue_link(cls.residues[i-1], res, "residue"):
+            length = i - start
+            if length > 1:
+                chains[alphabet] = {j + 1: start + j for j in range(length)}
+                chain_ids[start: start + length] = alphabet * length
+                alphabet = chr(ord(alphabet) + 1)
+            start = i
+    length = i - start
+    if length > 1:
+        chains[alphabet] = {j + 1: start + j for j in range(length)}
+        chain_ids[start: start + length] = alphabet * length
+    return chains, chain_ids
+
+
+def _pdb_connection(connects):
+    """
+
+    :param connects:
+    :return:
+    """
+    templist = []
+    for connect, atoms in connects.items():
+        atoms.sort()
+        atom_groups = [atoms[i:i + 4] for i in range(0, len(atoms), 4)]
+        for four_atoms in atom_groups:
+            templist.append("CONECT" + "{:5d}".format(connect + 1)
+                            + "".join(["{:5d}".format(i + 1) for i in four_atoms]) + "\n")
+    templist.sort()
+    return "".join(templist)
+
+
+def _pdb_sequence(cls: Molecule, chains: Xdict):
+    """
+
+    :param cls:
+    :param towrite:
+    :return:
+    """
+    towrite = ""
+    chain_ids = list(chains.keys())
+    chain_ids.sort()
+    for chain_id in chain_ids:
+        index_map = chains[chain_id]
+        pdb_index = list(index_map.keys())
+        pdb_index.sort()
+        names = []
+        for i in pdb_index:
+            mol_index = index_map[i]
+            name = cls.residues[mol_index].name
+            name = GlobalSetting.PDBResidueNameMap["save"][name] \
+                if name in GlobalSetting.PDBResidueNameMap["save"] else name
+            if len(name) > 3:
+                warn(f"The residue name {name} is more than 3 characters.")
+            names.append(name)
+        lines = [" ".join(["{:3s}".format(name) for name in names[j:j+13]])  for j in range(0, len(names), 13)]
+        lines = ["SEQRES {0:3d} {1:1s} {2:4d}  {3:s}\n".format(
+            i + 1, chain_id, len(names), line) for i, line in enumerate(lines)]
+        towrite += "".join(lines)
+    return towrite
+
+
[…]
