### 2022-09-06 cfd22ec
fix pdb problems (#22)
* update docs

* update the interface for the newest updated MindSPONGE

* update the functions to fix pdb problems
