+++ Xponge/helper/__init__.py
@@ -11,6 +11,7 @@ import stat
 from types import MethodType, FunctionType
 from functools import partial, partialmethod, wraps
 from collections import OrderedDict
+from collections.abc import Iterable
 from itertools import product
 from abc import ABC
 
@@ -20,16 +21,22 @@ from .namespace import set_real_global_variable, remove_real_global_variable, se
     set_attribute_alternative_name, set_classmethod_alternative_names, set_attribute_alternative_names, \
     set_dict_value_alternative_name, set_global_alternative_names, source
 
-from .math import get_rotate_matrix, get_fibonacci_grid, guess_element_from_mass
+from .math import get_rotate_matrix, get_fibonacci_grid, guess_element_from_mass, kabsch
 
 
 class Xdict(dict):
     """
     This **class** is used to be a dict which can give not_found_message
 
+    :param not_found_method: the method (function) which will accept the key and return the value \
+when the key is not found. **New From 1.2.6.7**
     :param not_found_message: the string to print when the key is not found
     """
     def __init__(self, *args, **kwargs):
+        if "not_found_method" in kwargs:
+            self.not_found_method = kwargs.pop("not_found_method")
+        else:
+            self.not_found_method = None
         if "not_found_message" in kwargs:
             self.not_found_message = kwargs.pop("not_found_message")
         else:
@@ -41,6 +48,10 @@ class Xdict(dict):
         toget = self.get(key, self.id)
         if toget != self.id:
             return toget
+        if self.not_found_method:
+            value = self.not_found_method(key)
+            self[key] = value
+            return self[key]
         if self.not_found_message:
             raise KeyError(self.not_found_message.format(key))
         raise KeyError(f"{key}")
@@ -959,6 +970,41 @@ class Residue(Entity):
             return np.sum([getattr(atom, attr) for atom in self.atoms])
         return super().__getattribute__(attr)
 
+    def unterminal(self, add_missing_atoms=False):
+        """
+        This **function** is used to turn the terminal residue to be unterminal
+        **New From 1.2.6.7**
+
+        :param add_missing_atoms: whether to add missing atoms after deleting the terminal atoms
+        :return: 1 for success, 0 for failure
+        """
+        if self.type.name in GlobalSetting.PDBResidueNameMap["save"]:
+            new_type = ResidueType.get_type(GlobalSetting.PDBResidueNameMap["save"][self.type.name])
+        else:
+            return 0
+        new_type_atom = Xdict({atom.name: atom for atom in new_type.atoms})
+        to_remove = set()
+        for atom in self.atoms:
+            if atom.name not in new_type_atom:
+                to_remove.add(atom)
+            else:
+                atom.type = new_type_atom[atom.name].type
+                atom.charge = new_type_atom[atom.name].charge
+                self._name2index[atom.name] -= len(to_remove)
+                self._atom2index[atom] -= len(to_remove)
+        for atom in to_remove:
+            self._name2index.pop(atom.name)
+            self._atom2index.pop(atom)
+            self._atom2name.pop(atom)
+            self._name2atom.pop(atom.name)
+            self.atoms.remove(atom)
+            atom.residue = None
+        self.type = new_type
+        self.name = new_type.name
+        if add_missing_atoms:
+            self.add_missing_atoms()
+        return 1
+
     def name2atom(self, name):
         """
         This **function** convert an atom name to an Atom instance
@@ -1061,19 +1107,30 @@ class Residue(Entity):
         """
         t = {atom.name for atom in self.atoms}
         uncertified = {atom.name for atom in self.type.atoms}
+        certified_positions = []
+        template_positions = []
         for atom in self.type.atoms:
             if atom.name in t:
[…]
