### 2022-09-06 87c6ff6
reduce cyclomatic complexity, correct code spelling mistakes and pylint warnings
