### 2022-09-06 ab2d8b6
reduce cyclomatic complexity, correct code spelling mistakes and pylint warnings (#23)
Co-authored-by: pkuxyj <yijiexia@pku.edu.cn>
