### 2022-09-06 7a9295d
SpongeTrajectoryReader fits for md analysis chain reader
