### 2022-09-06 c001f93
add the function to omit atoms and update the docs
