+++ Xponge/helper/__init__.py
@@ -980,6 +980,7 @@ class Residue(Entity):
         self.atoms = []
         """the atoms in the instance"""
         self._name2atom = Xdict()
+        self._name2atom.not_found_message = "There is no atom named {} in %s" % self.name
         self._atom2name = Xdict()
         self._atom2index = Xdict()
         self._name2index = Xdict()
@@ -1007,18 +1008,17 @@ class Residue(Entity):
             return np.sum([getattr(atom, attr) for atom in self.atoms])
         return super().__getattribute__(attr)
 
-    def unterminal(self, add_missing_atoms=False):
+    def set_type(self, new_type, add_missing_atoms=False):
         """
-        This **function** is used to turn the terminal residue to be unterminal
-        **New From 1.2.6.7**
+        This **function** is used to change the type of the residue to a new type
+        **New From 1.2.6.8**
 
+        :param new_type: the instance or the name of the new residue type
         :param add_missing_atoms: whether to add missing atoms after deleting the terminal atoms
         :return: 1 for success, 0 for failure
         """
-        if self.type.name in GlobalSetting.PDBResidueNameMap["save"]:
-            new_type = ResidueType.get_type(GlobalSetting.PDBResidueNameMap["save"][self.type.name])
-        else:
-            return 0
+        if isinstance(new_type, str):
+            new_type = ResidueType.get_type(new_type)
         new_type_atom = Xdict({atom.name: atom for atom in new_type.atoms})
         to_remove = set()
         for atom in self.atoms:
@@ -1041,6 +1041,20 @@ class Residue(Entity):
         if add_missing_atoms:
             self.add_missing_atoms()
         return 1
+
+    def unterminal(self, add_missing_atoms=False):
+        """
+        This **function** is used to turn the terminal residue to be unterminal
+        **New From 1.2.6.7**
+
+        :param add_missing_atoms: whether to add missing atoms after deleting the terminal atoms
+        :return: 1 for success, 0 for failure
+        """
+        if self.type.name in GlobalSetting.PDBResidueNameMap["save"]:
+            new_type = GlobalSetting.PDBResidueNameMap["save"][self.type.name]
+        else:
+            return 0
+        return self.set_type(new_type)
 
     def name2atom(self, name):
         """
@@ -1855,18 +1869,18 @@ def _link_residue_process_coordinate(molecule, atom1, atom2):
                                          get_rotate_matrix(np.cross(r_ao, r_ob), delta_angle)) + \
                                          crd[molecule.atom_index[atoms[1]]]
         elif len(atoms) == 3:
-            r_oo = crd[molecule.atom_index[atoms[0]]] - crd[molecule.atom_index[atoms[1]]]
-            r_oa = crd[molecule.atom_index[atoms[1]]] - crd[molecule.atom_index[atoms[2]]]
-            r_ab = crd[molecule.atom_index[atoms[1]]] - crd[molecule.atom_index[atom_b]]
-            r12xr23 = np.cross(r_oo, r_oa)
-            r23xr34 = np.cross(r_ab, r_oa)
-            cos = np.dot(r12xr23, r23xr34) / np.linalg.norm(r12xr23) / np.linalg.norm(r23xr34)
+            r12 = crd[molecule.atom_index[atoms[0]]] - crd[molecule.atom_index[atoms[1]]]
[…]
