+++ Xponge/forcefield/amber/tip4pew.py
@@ -1,7 +1,7 @@
 """
 This **module** set the basic configuration for ff14sb
 """
-from ...helper import source, Xprint, set_real_global_variable
+from ...helper import source, Xprint
 
 source("....")
 amber = source("...amber")
@@ -38,17 +38,15 @@ name   atom0    atom1   atom2   k1         k2
 EP     -3       -2      -1      0.1066413  0.1066413
 """)
 
-TIP4P = load_mol2(os.path.join(AMBER_DATA_DIR, "tip4pew.mol2"))
+TIP4P = load_mol2(os.path.join(AMBER_DATA_DIR, "tip4pew.mol2"), as_template=True)
 
 amber.load_parameters_from_frcmod("ions1lm_126_tip4pew.frcmod")
 amber.load_parameters_from_frcmod("ionsjc_tip4pew.frcmod")
 amber.load_parameters_from_frcmod("ions234lm_126_tip4pew.frcmod")
 
-load_mol2(os.path.join(AMBER_DATA_DIR, "atomic_ions.mol2"))
+load_mol2(os.path.join(AMBER_DATA_DIR, "atomic_ions.mol2"), as_template=True)
 
 set_real_global_variable("WAT", TIP4P)
-
-load_mol2(os.path.join(AMBER_DATA_DIR, "atomic_ions.mol2"))
 
 Xprint("""Reference for tip4pew:
 1. Water:

[…]
