### 2022-09-11 Xia-Yijie/Xponge#4 @ojeda-e
Thanks @Xia-Yijie, is it possible to specify what is not available on Windows? A note in the docs will be sufficient. It's better to be clear with the user about the limitations when working on Windows. Once that is cleared out I'll close this issue.
