### 2022-09-11 #4467 @hmacdope
@Xia-Yijie congratulations. It is recommended that you make a new git tag and archive it on Zenodo or similar. See the docs [here](https://joss.readthedocs.io/en/latest/editing.html).

Other than that @richardjgowers we are good to roll. 👍👍
