+++ docs/source/Xponge.forcefield.amber.glycam_06j.rst
@@ -1,0 +1,50 @@
+Xponge.forcefield.amber.glycam\_06j package
+===========================================
+
+.. automodule:: Xponge.forcefield.amber.glycam_06j
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+Xponge.forcefield.amber.glycam\_06j.d\_furanose module
+------------------------------------------------------
+
+.. automodule:: Xponge.forcefield.amber.glycam_06j.d_furanose
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.amber.glycam\_06j.d\_pyranose module
+------------------------------------------------------
+
+.. automodule:: Xponge.forcefield.amber.glycam_06j.d_pyranose
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.amber.glycam\_06j.glycoprotein module
+-------------------------------------------------------
+
+.. automodule:: Xponge.forcefield.amber.glycam_06j.glycoprotein
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.amber.glycam\_06j.l\_furanose module
+------------------------------------------------------
+
+.. automodule:: Xponge.forcefield.amber.glycam_06j.l_furanose
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Xponge.forcefield.amber.glycam\_06j.l\_pyranose module
+------------------------------------------------------
+
+.. automodule:: Xponge.forcefield.amber.glycam_06j.l_pyranose
+   :members:
+   :undoc-members:
+   :show-inheritance:
