### 2022-09-13 786de54
update to 1.2.6.8 (#24)
* reduce cyclomatic complexity, correct code spelling mistakes and pylint warnings

* SpongeTrajectoryReader fits for md analysis chain reader

* add the function to omit atoms and update the docs

* new forcefield: glycam

* new force field for DNA and RNA

* update the docs
