+++ Xponge/load.py
@@ -3,6 +3,7 @@ This **module** is used to load and read
 """
 import os
 import warnings
+import re
 
 import numpy as np
 
@@ -82,14 +83,102 @@ def _mol2_bond(line, current_molecule, atom_residue_map):
                     atom_residue_map[words[2]][1].type.tail = atom_residue_map[words[2]][0]
 
 
-def load_mol2(filename, ignore_atom_type=False):
+def _mol2_template_atom(line, current_residue_index, temp, current_residue, atom_residue_map):
+    """
+
+    :param line:
+    :param current_residue_index:
+    :param current_residue:
+    :param ignore_atom_type:
+    :param temp:
+    :param current_molecule:
+    :param atom_residue_map:
+    :return:
+    """
+    words = line.split()
+    if current_residue_index is None or int(words[6]) != current_residue_index:
+        current_residue_index = int(words[6])
+        if words[7] not in ResidueType.get_all_types():
+            set_real_global_variable(words[7], ResidueType(name=words[7]))
+            temp = True
+        else:
+            temp = False
+        current_residue = ResidueType.get_type(words[7])
+
+    temp_atom_type = AtomType.get_type(words[5])
+
+    if temp:
+        current_residue.Add_Atom(words[1], temp_atom_type, *words[2:5])
+        current_residue.atoms[-1].Update(**{"charge[e]": float(words[8])})
+    atom_residue_map[words[0]] = [words[1], current_residue, current_residue_index, temp]
+    return current_residue_index, current_residue, temp
+
+
+def _mol2_template_bond(line, atom_residue_map):
+    """
+
+    :param line:
+    :param current_molecule:
+    :param atom_residue_map:
+    :return:
+    """
+    words = line.split()
+    atom1_index, atom2_index = words[1], words[2]
+    atom_name, residue_type, residue_index, is_new = 0, 1, 2, 3
+    atom1_info, atom2_info = atom_residue_map[atom1_index], atom_residue_map[atom2_index]
+    if atom1_info[residue_type] == atom2_info[residue_type] and atom1_info[is_new]:
+        atom1_info[residue_type].Add_Connectivity(atom1_info[atom_name], atom2_info[atom_name])
+    else:
+        index_diff = atom1_info[residue_index] - atom2_info[residue_index]
+        if abs(index_diff) == 1:
+            if atom1_info[is_new]:
+                if index_diff < 0:
+                    atom1_info[residue_type].tail = atom1_info[atom_name]
+                else:
+                    atom1_info[residue_type].head = atom1_info[atom_name]
+            if atom2_info[is_new]:
+                if index_diff < 0:
+                    atom2_info[residue_type].head = atom2_info[atom_name]
+                else:
+                    atom2_info[residue_type].tail = atom2_info[atom_name]
+
+
+def _mol2_template(filename):
+    """
+    This is used to make loading more efficient when importing the force fields
+
+    :param filename:
+    :return:
+    """
+    current_residue_index = None
+    current_residue = None
+    atom_residue_map = Xdict()
+    temp = None
+    with open(filename) as f:
+        flag = None
+        for line in f:
+            if not line.strip():
+                continue
+            if line.startswith("@<TRIPOS>"):
+                flag = line[9:].strip()
+            elif flag == "ATOM":
+                current_residue_index, current_residue, temp = _mol2_template_atom(line, current_residue_index, temp,
+                                                                                   current_residue, atom_residue_map)
+            elif flag == "BOND":
[…]
