### 2022-09-13 Xia-Yijie/Xponge#4 @Xia-Yijie
@ojeda-e now a link to [the detailed list](https://spongemm.cn/xponge_doc/dependency.html#unavalable-functions-on-windows) has been added to [the first paragraph of installation in README](https://github.com/Xia-Yijie/Xponge/blob/master/README.md#installation) and the hints have been added to the API documentation of the related functions, too.
