### 2022-09-20 #4467 @kyleniemeyer
@Xia-Yijie I think the bot struggled with those items; can you confirm those are the appropriate citations for those things? In particular, for the software, are there DOIs available?
