### 2022-09-21 #4467 @Xia-Yijie
@kyleniemeyer I find that there are three available DOIs for the software and I have added them to the file `paper.bib`, but there are no available DOIs for the others.
Every time I asked the bot to check references, the bot gave me `None` in `MISSING DOIs`, but when you asked it to check references, it raised some errors. Why?
