### 2022-09-22 #4467 @Xia-Yijie
> @Xia-Yijie Thanks for adding those. It could have been a temporary issue with something behind the scenes. At the end of the day, those messages are meant more as guidance anyways.

@kyleniemeyer  Thanks for you reply. I see.

> 👉📄 [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04467/joss.04467/10.21105.joss.04467.pdf) 📄 [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.04467/joss.04467/10.21105.joss.04467.pdf) 📄 👈

The final proof looks good to me.
