+++ paper.bib
@@ -202,18 +202,26 @@ year = {2011}
     url = {https://ambermd.org/tutorials/pengfei/index.php}  
 }
 
-@misc{CGenFF,
-    author = {Alex Mackerell and Kenno Vanommeslaeghe},
-    year = {2022},
-    title = {CGenFF Home},
-    url = {https://cgenff.umaryland.edu/}  
+@article{CGenFF,
+author = {Vanommeslaeghe, K. and Raman, E. Prabhu and MacKerell, A. D.},
+title = {Automation of the CHARMM General Force Field (CGenFF) II: Assignment of Bonded Parameters and Partial Atomic Charges},
+journal = {Journal of Chemical Information and Modeling},
+volume = {52},
+number = {12},
+pages = {3155-3168},
+year = {2012},
+doi = {10.1021/ci3003649},
+note ={PMID: 23145473},
+URL = {https://doi.org/10.1021/ci3003649},
+eprint = {https://doi.org/10.1021/ci3003649}
 }
 
 @misc{pdb2gmx,
     author = {Paul Bauer; Berk Hess; Erik Lindahl},
     year = {2022},
     title = {gmx pdb2gmx},
-    url = {https://manual.gromacs.org/documentation/current/onlinehelp/gmx-pdb2gmx.html}  
+    url = {https://manual.gromacs.org/documentation/current/onlinehelp/gmx-pdb2gmx.html},
+    doi = { 10.5281/zenodo.7037337 }
 }
 
 @misc{psfgen,
@@ -234,7 +242,8 @@ year = {2011}
     author = {Greg Landrum, Paolo Tosco, Brian Kelley, Ric, sriniker, gedeck, Riccardo Vianello, NadineSchneider, Eisuke Kawashima, David Cosgrove, Andrew Dalke, Dan N, Gareth Jones, Brian Cole, Matt Swain, Samo Turk, AlexanderSavelyev, Alain Vaucher, Maciej Wójcikowski, Ichiru Take, Daniel Probst, Kazuya Ujihara, Vincent F. Scalfani, guillaume godin, Axel Pahl, Francois Berenger, JLVarjo, strets123, JP, DoliathGavid},
     year = {2022},
     title = {rdkit/rdkit: 2022_03_5 (Q1 2022) Release},
-    url = {https://zenodo.org/record/6961488#.YvackPhByUk}  
+    url = {https://zenodo.org/record/6961488#.YvackPhByUk},
+    doi = {10.5281/zenodo.6961488}
 }
 
 @misc{MindSpore,
